refactor(server): split conn.go into cohesive same-package files (phase 1 of #722)
Pure mechanical move of top-level declarations out of server/conn.go
(6,887 lines) into seven cohesive files in the same package:

- conn_errors.go            error classification / SQLSTATE mapping
- conn_query_exec.go        simple-protocol query execution paths
- conn_results.go           row description / DataRow encoding / value formatting
- conn_copy.go              COPY IN/OUT protocol (text/CSV/binary/remote streaming)
- conn_extended_query.go    Parse/Describe/Execute/Close handlers
- conn_pg_stat_activity.go  pg_stat_activity emulation
- conn_cursor.go            server-side cursor + pg_cursors emulation

Zero renames, zero signature changes, zero logic edits — verified by
parsing origin/main's conn.go and the new file set and asserting the
multisets of go/printer-normalized top-level declarations are identical.

Deliberately left in conn.go to avoid conflicting with in-flight PRs:
handleBind, the messageLoop dispatch, and the startup/cleartext-auth
section. isDMLReturning and friends also stay in conn.go (referenced
there by CLAUDE.md).

Partially addresses #722 (phase 1 of 3).
