+++ server/conn.go
@@ -7,7 +7,6 @@ import (
 	"crypto/tls"
 	"database/sql"
 	"encoding/binary"
-	"encoding/csv"
 	"errors"
 	"fmt"
 	"io"
@@ -15,18 +14,14 @@ import (
 	"net"
 	"os"
 	"regexp"
-	"sort"
 	"strconv"
 	"strings"
 	"sync/atomic"
 	"time"
 
 	pg_query "github.com/pganalyze/pg_query_go/v6"
-	"github.com/posthog/duckgres/duckdbservice/arrowmap"
 	"github.com/posthog/duckgres/server/auth"
-	"github.com/posthog/duckgres/server/flightclient"
 	"github.com/posthog/duckgres/server/iceberg"
-	"github.com/posthog/duckgres/server/icebergmeta"
 	"github.com/posthog/duckgres/server/observe"
 	"github.com/posthog/duckgres/server/sessionmeta"
 	"github.com/posthog/duckgres/server/sqlcore"
@@ -375,298 +370,6 @@ func (c *clientConn) startDisconnectMonitor(ctx context.Context) (stop func()) {
 	}
 }
 
-// isQueryCancelled checks whether an error string indicates that *some*
-// context cancellation was involved. It does NOT distinguish caller-driven
-// cancellation (Ctrl-C, client disconnect, deadline) from infra-driven
-// cancellation (worker died, CP closed the gRPC client) — both surface here
-// as a wrapped "context canceled" string. Use it for SQLSTATE classification
-// where 57014 is correct either way; do not use it to gate error logging,
-// since infra cancels are real failures we want surfaced. Use
-// (*clientConn).isCallerCancellation for that.
-func isQueryCancelled(err error) bool {
-	return err == context.Canceled || (err != nil && strings.Contains(err.Error(), "context canceled"))
-}
-
-// isCallerCancellation reports whether err is a cancellation that the caller
-// asked for — either through pgwire CancelRequest, an explicit ctx cancel, or
-// a deadline. Distinct from a gRPC Canceled status that bubbles up purely
-// because the CP closed the worker connection (worker death, takeover): in
-// that case c.ctx is still healthy, c.ctx.Err() is nil, and the surface error
-// must be logged as an infra failure rather than suppressed as "user
-// cancelled". This matters for alerting — "Query execution errored." should
-// fire on worker kills, not get silently downgraded to "Query finished.".
-func (c *clientConn) isCallerCancellation(err error) bool {
-	if !isQueryCancelled(err) {
-		return false
-	}
-	if c == nil || c.ctx == nil {
-		return false
-	}
-	return c.ctx.Err() != nil
-}
-
-// isDuckLakeTransactionConflict returns true if the error is a DuckLake
-// transaction conflict. These occur when concurrent DuckLake transactions
-// try to commit overlapping changes. DuckLake uses global snapshot IDs, so
-// even writes to unrelated tables can conflict under concurrency.
-func isDuckLakeTransactionConflict(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := err.Error()
-	return strings.Contains(msg, "Transaction conflict")
-}
-
-// isDuckLakeMetadataConnectionLost returns true if the error indicates the
-// DuckLake metadata store connection was lost during a transaction. This
-// typically happens when long-running queries leave the metadata connection
-// idle long enough for RDS or a network layer to drop it.
-func isDuckLakeMetadataConnectionLost(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := err.Error()
-	return strings.Contains(msg, "DuckLake transaction") &&
-		strings.Contains(msg, "SSL connection has been closed unexpectedly")
-}
-
-// classifyErrorCode returns the most appropriate PostgreSQL SQLSTATE for a
-// DuckDB error. Transaction conflicts get 40001 (serialization_failure), which
-// signals PG-aware clients to retry. Query cancellations get 57014. Remaining
-// errors are classified by DuckDB's exception-type prefix; the prefix is the
-// only signal the Go driver exposes today, so we parse the message string.
-func classifyErrorCode(err error) string {
-	// Wire-side SQLSTATE: 57014 is correct for any cancellation regardless of
-	// whether the caller drove it. Use isQueryCancelled here, not the
-	// clientConn-aware variant — classifyErrorCode is called from contexts
-	// that don't have a *clientConn (e.g., direct test fixtures).
-	if isQueryCancelled(err) {
-		return "57014"
-	}
-	if isDuckLakeTransactionConflict(err) {
-		return "40001" // serialization_failure — client should retry
-	}
-
-	// Workers reach the control plane over Arrow Flight SQL, so DuckDB errors
-	// arrive wrapped as "flight execute update: rpc error: code = X desc = <msg>".
-	// Unwrap to recover the underlying DuckDB exception message so the prefix
-	// classifiers below apply; otherwise every Iceberg/DuckLake worker error
-	// would fall through to XX000 and the client would see the raw rpc string.
-	msg := unwrapFlightError(err.Error())
-	// Iceberg schema-evolution failure: after a column is dropped (or other
-	// schema churn) DuckDB's iceberg extension can fail every scan with
-	// "Tried to scan a snapshot created with a newer schema id ...". Surface it
-	// as feature_not_supported with an actionable message rather than XX000.
-	if isSchemaEvolutionErrorMsg(msg) {
-		return "0A000"
-	}
-	switch {
-	case strings.HasPrefix(msg, "Catalog Error:"):
-		return catalogErrorCode(msg)
-	case strings.HasPrefix(msg, "Binder Error:"):
-		return binderErrorCode(msg)
-	case strings.HasPrefix(msg, "Parser Error:"):
-		return "42601" // syntax_error
-	case strings.HasPrefix(msg, "Conversion Error:"):
-		return conversionErrorCode(msg)
-	case strings.HasPrefix(msg, "Out of Range Error:"):
-		return "22003" // numeric_value_out_of_range
-	case strings.HasPrefix(msg, "Constraint Error:"):
-		return constraintErrorCode(msg)
-	case strings.HasPrefix(msg, "Permission Error:"):
-		return "42501" // insufficient_privilege
-	case strings.HasPrefix(msg, "Not implemented Error:"),
-		strings.HasPrefix(msg, "Invalid Input Error: Not implemented"):
-		return "0A000" // feature_not_supported
-	case strings.HasPrefix(msg, "Transaction Error:"),
-		strings.HasPrefix(msg, "TransactionContext Error:"):
-		return "25000" // invalid_transaction_state — DuckDB emits both prefixes
-	case strings.HasPrefix(msg, "Dependency Error:"):
-		return "2BP01" // dependent_objects_still_exist
-	}
-	// Unknown error class — no DuckDB prefix matched. These are
-	// typically infra issues (gRPC failures, IO errors, internal panics)
-	// rather than user input issues. Classify as XX000 (internal_error)
-	// so isUserQueryError correctly routes them to the system-error log
-	// path. If a future DuckDB error needs to land in a user class, add
-	// a prefix branch above instead of moving the fallback.
-	return "XX000"
-}
-
-// transformErrorSQLState returns the SQLSTATE for a transpiler-detected error.
-// A transform may attach an explicit code (e.g. 0A000 for an unsupported
-// feature via transform.CodedError); otherwise it defaults to 42704
-// (undefined_object), matching the historical behavior for unrecognized config
-// parameters.
-func transformErrorSQLState(err error) string {
-	var coded interface{ SQLState() string }
-	if errors.As(err, &coded) {
-		return coded.SQLState()
-	}
-	return "42704"
-}
-
-// isSchemaEvolutionErrorMsg reports whether an (already Flight-unwrapped) DuckDB
-// error message is the Iceberg "newer schema id" scan failure that a DROP COLUMN
-// can leave a table in. Matched on a stable substring; if the wording changes
-// the error simply falls back to its generic class (still safe).
-func isSchemaEvolutionErrorMsg(msg string) bool {
-	return strings.Contains(msg, "newer schema id")
-}
-
-// friendlyExecError rewrites known-cryptic DuckDB/Iceberg execution errors into
-// actionable client messages. Returns the original message unchanged otherwise.
-func friendlyExecError(err error) string {
-	msg := err.Error()
-	if isSchemaEvolutionErrorMsg(unwrapFlightError(msg)) {
-		return "reading this table failed because a column was dropped: DROP COLUMN is not safely supported on this catalog (it can leave the table unreadable after schema changes); recreate the table instead"
-	}
-	return msg
-}
-
-// unwrapFlightError recovers the underlying error message from an Arrow Flight /
-// gRPC wrapper. gRPC errors stringify as "… rpc error: code = X desc = <message>",
-// and the control plane further prefixes worker failures with "flight execute
-// update: " (and similar). When the message is not Flight-wrapped it is returned
-// unchanged, so this is safe to apply unconditionally before prefix matching.
-func unwrapFlightError(msg string) string {
-	// gRPC puts the real message after the final "desc = ".
-	if idx := strings.LastIndex(msg, "desc = "); idx != -1 {
-		return strings.TrimSpace(msg[idx+len("desc = "):])
-	}
-	return strings.TrimSpace(msg)
-}
-
-// catalogErrorCode narrows a "Catalog Error: …" message to a specific SQLSTATE
-func catalogErrorCode(msg string) string {
-	lower := strings.ToLower(msg)
-	switch {
-	case strings.Contains(lower, "schema") && strings.Contains(lower, "does not exist"):
-		return "3F000" // invalid_schema_name
-	case strings.Contains(lower, "table with name") && strings.Contains(lower, "does not exist"):
-		return "42P01" // undefined_table
-	case strings.Contains(lower, "view with name") && strings.Contains(lower, "does not exist"):
-		return "42P01" // undefined_table (views share the code)
-	case strings.Contains(lower, "function") && (strings.Contains(lower, "does not exist") || strings.Contains(lower, "with these arguments")):
-		return "42883" // undefined_function
-	case strings.Contains(lower, "no function matches"):
-		return "42883" // undefined_function — DuckDB's overload-resolution failure
-	case strings.Contains(lower, "type") && strings.Contains(lower, "does not exist"):
-		return "42704" // undefined_object
-	case strings.Contains(lower, "does not exist"):
-		return "42704" // undefined_object (generic fallback)
-	case strings.Contains(lower, "already exists"):
-		if strings.Contains(lower, "function") {
-			return "42723" // duplicate_function
-		}
-		if strings.Contains(lower, "schema") {
-			return "42P06" // duplicate_schema
-		}
-		return "42P07" // duplicate_table
-	}
-	return "42000"
-}
-
-// binderErrorCode narrows a "Binder Error: …" message. The binder raises on
-// semantic problems discovered after parsing, most commonly missing columns.
-func binderErrorCode(msg string) string {
-	lower := strings.ToLower(msg)
-	switch {
-	case strings.Contains(lower, "referenced column") && strings.Contains(lower, "not found"):
-		return "42703" // undefined_column
-	case strings.Contains(lower, "column") && strings.Contains(lower, "does not exist"):
-		return "42703"
-	case strings.Contains(lower, "ambiguous"):
-		return "42702" // ambiguous_column
-	case strings.Contains(lower, "referenced table") && strings.Contains(lower, "not found"):
-		return "42P01" // undefined_table — DuckDB raises this for unknown aliases
-	case strings.Contains(lower, "no function matches"):
-		return "42883" // undefined_function — overload-resolution failure
-	}
-	return "42601" // syntax_error — binder failures without a narrower match
-}
-
-// conversionErrorCode narrows a "Conversion Error: …" message. DuckDB uses
-// this prefix for both invalid text representations and numeric overflows
-// during casts (e.g. CAST(1000 AS TINYINT));
-func conversionErrorCode(msg string) string {
-	lower := strings.ToLower(msg)
-	switch {
-	case strings.Contains(lower, "out of range"),
-		strings.Contains(lower, "overflow"),
-		strings.Contains(lower, "would be out of range"):
-		return "22003" // numeric_value_out_of_range
-	}
-	return "22P02" // invalid_text_representation
-}
-
-// constraintErrorCode narrows a "Constraint Error: …" message to one of the
-// integrity_constraint_violation family codes. DuckDB's messages name the
-// violated constraint explicitly, so substring matching is reliable.
-func constraintErrorCode(msg string) string {
-	lower := strings.ToLower(msg)
-	switch {
-	case strings.Contains(lower, "duplicate key") || strings.Contains(lower, "unique"):
-		return "23505" // unique_violation
-	case strings.Contains(lower, "not null") || strings.Contains(lower, "null value"):
-		return "23502" // not_null_violation
-	case strings.Contains(lower, "foreign key"):
-		return "23503" // foreign_key_violation
-	case strings.Contains(lower, "check constraint"):
-		return "23514" // check_violation
-	}
-	return "23000" // integrity_constraint_violation
-}
-
-// userErrorSQLSTATEClasses is the closed set of PostgreSQL SQLSTATE class
-// codes (the first two characters) that represent user-input or
-// user-state errors — "you wrote a query that doesn't make sense for
-// this database state." Anything outside this set is treated as a
-// system / infra error (08 connection, 53 resources, 57 operator
-// intervention, 58 system, XX internal, …).
-//
-// The discriminator is the SQLSTATE we already compute for the pgwire
-// error response — no new string matching here. Add a class only after
-// confirming every code in it is genuinely user-attributable; adding
-// erroneously will hide real infra failures from the alert path.
-var userErrorSQLSTATEClasses = map[string]struct{}{
-	"0A": {}, // feature_not_supported — user used a SQL feature we don't have
-	"22": {}, // data_exception — bad input (cast, overflow, encoding)
-	"23": {}, // integrity_constraint_violation — unique/fk/check/not_null
-	"25": {}, // invalid_transaction_state — nested BEGIN, etc.
-	"28": {}, // invalid_authorization_specification — not hit on this path today
-	"2B": {}, // dependent_objects_still_exist — DROP without CASCADE
-	"3D": {}, // invalid_catalog_name — DB doesn't exist
-	"3F": {}, // invalid_schema_name — schema doesn't exist
-	"42": {}, // syntax_error_or_access_rule_violation — table/column not found, syntax
-	"44": {}, // with_check_option_violation
-}
-
-// isUserQueryError tells the log/observability path whether a query
-// failure is user-attributable (Info-level "Query execution failed.")
-// or a real system error worth alerting on (Error-level "Query
-// execution errored."). The discriminator is the SQLSTATE class —
-// already computed via classifyErrorCode for the pgwire response.
-//
-// Cancellations (SQLSTATE 57014) are NOT short-circuited as user errors
-// here. The call sites (logQueryError, the clientConn error paths) gate
-// caller-driven cancellations upstream via clientConn.isCallerCancellation,
-// so by the time isUserQueryError sees a cancel-shaped err the cancellation
-// was infra-driven (gRPC client closed because the worker died, the worker
-// was retired, etc.) and class 57 = operator intervention is the right
-// bucket — that's a real failure we want surfaced at Error level.
-func isUserQueryError(err error) bool {
-	if err == nil {
-		return false
-	}
-	code := classifyErrorCode(err)
-	if len(code) < 2 {
-		return false
-	}
-	_, ok := userErrorSQLSTATEClasses[code[:2]]
-	return ok
-}
-
 // logQueryStarted records a query handing off to a worker. Pairs with
 // logQueryFinished at every termination point so logs and traces can
 // be cross-referenced — the trace_id attribute matches the OTEL span
@@ -732,23 +435,6 @@ func (c *clientConn) logQueryError(query string, err error) {
 		return
 	}
 	slog.Error("Query execution errored.", attrs...)
-}
-
-// isConnectionBroken checks if an error indicates a broken connection
-// (e.g., SSL connection closed, network error).
-func isConnectionBroken(err error) bool {
-	if err == nil {
-		return false
-	}
-	errMsg := strings.ToLower(err.Error())
-	return strings.Contains(errMsg, "ssl connection has been closed") ||
-		strings.Contains(errMsg, "connection refused") ||
-		strings.Contains(errMsg, "broken pipe") ||
-		strings.Contains(errMsg, "connection reset") ||
-		strings.Contains(errMsg, "network is unreachable") ||
-		strings.Contains(errMsg, "no route to host") ||
-		strings.Contains(errMsg, "i/o timeout") ||
-		strings.Contains(errMsg, "use of closed network connection")
 }
 
 // safeCleanupDB safely closes the database connection, handling the case where
@@ -1690,1194 +1376,6 @@ func (c *clientConn) handleQuery(body []byte) error {
 	return err
 }
 
-// executeQueryDirect executes a query directly against DuckDB without any transpilation.
-// Used for passthrough users who send DuckDB-native SQL.
-func (c *clientConn) executeQueryDirect(query, cmdType string) error {
-	if !queryReturnsResults(query) {
-		// Handle nested BEGIN
-		if cmdType == "BEGIN" && c.txStatus == txStatusTransaction {
-			c.sendNotice("WARNING", "25001", "there is already a transaction in progress")
-			_ = wire.WriteCommandComplete(c.writer, "BEGIN")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-
-		ctx, cleanup := c.queryContext()
-		defer cleanup()
-
-		// Lifecycle log pair (PR #519): every DML simple-query gets a
-		// matched logQueryStarted / logQueryFinished. Captured via
-		// closures so the deferred call sees the eventual rows + err.
-		queryStart := time.Now()
-		var queryRowsAff int64
-		var queryFinalErr error
-		c.logQueryStarted(query)
-		defer func() {
-			c.logQueryFinished(query, queryStart, queryRowsAff, queryFinalErr)
-		}()
-
-		runExec := func() (ExecResult, error) {
-			return c.executor.ExecContext(ctx, query)
-		}
-
-		result, err := runExec()
-		if err != nil && c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-			ducklakeConflictTotal.Inc()
-			result, err = retryOnConflict(runExec)
-		}
-		if err != nil {
-			result, err, _ = recoverAbortedTransaction(
-				err,
-				c.txStatus == txStatusIdle,
-				func() error {
-					_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-					return rollbackErr
-				},
-				func() (ExecResult, error) {
-					return c.executor.ExecContext(ctx, query)
-				},
-			)
-		}
-		if err != nil {
-			queryFinalErr = err
-			errCode := classifyErrorCode(err)
-			errMsg := friendlyExecError(err)
-			if c.isCallerCancellation(err) {
-				errMsg = "canceling statement due to user request"
-			} else {
-				c.logQueryError(query, err)
-			}
-			c.sendError("ERROR", errCode, errMsg)
-			c.setTxError()
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-
-		if result != nil {
-			queryRowsAff, _ = result.RowsAffected()
-		}
-		c.updateTxStatus(cmdType)
-		tag := c.buildCommandTag(cmdType, result)
-		_ = wire.WriteCommandComplete(c.writer, tag)
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	_, _, _, err := c.executeSelectQuery(query, cmdType)
-	return err
-}
-
-// rewriteDirectQuery expands a bare `USE ducklake`/`USE iceberg` to its reliable
-// two-part `catalog.schema` target. This is NOT logical-name masking — the
-// catalog names are real; the rewrite only works around DuckDB's bare-catalog
-// `USE` resolution (a bare `USE ducklake` issued while the session is in the
-// iceberg catalog resolves `ducklake` as a *schema* within iceberg, landing on a
-// bogus `iceberg.ducklake`). Any other `USE <name>` and all other statements are
-// passed through unchanged.
-func (c *clientConn) rewriteDirectQuery(query string) string {
-	if c == nil || c.server == nil || c.passthrough || !c.catalogUseRewrite {
-		return query
-	}
-
-	stripped := strings.TrimSpace(stripLeadingComments(query))
-	if stripped == "" {
-		return query
-	}
-
-	hasSemicolon := strings.HasSuffix(stripped, ";")
-	trimmed := strings.TrimSpace(strings.TrimSuffix(stripped, ";"))
-
-	if len(trimmed) < len("USE") || !strings.EqualFold(trimmed[:len("USE")], "USE") {
-		return query
-	}
-
-	target := strings.TrimSpace(trimmed[len("USE"):])
-	if target == "" {
-		return query
-	}
-
-	unquoted := target
-	if len(target) >= 2 && target[0] == '"' && target[len(target)-1] == '"' {
-		unquoted = strings.ReplaceAll(target[1:len(target)-1], `""`, `"`)
-	}
-
-	var target2part string
-	switch {
-	case strings.EqualFold(unquoted, physicalDuckLakeCatalog):
-		// `USE ducklake` -> ducklake.main (DuckLake's real schema is `main`).
-		target2part = physicalDuckLakeCatalog + ".main"
-	case strings.EqualFold(unquoted, iceberg.CatalogName):
-		// `USE iceberg` -> iceberg.<DefaultSchema>. DuckDB can't `USE` a bare REST
-		// catalog (it targets <catalog>.main, which it shadows), so land on the
-		// guaranteed default schema (ensured by attachLakekeeperCatalog).
-		target2part = iceberg.CatalogName + "." + iceberg.DefaultSchema
-	default:
-		return query
-	}
-
-	rewritten := "USE " + target2part
-	if hasSemicolon {
-		rewritten += ";"
-	}
-	return rewritten
-}
-
-func (c *clientConn) loadIcebergColumnMetadata(ctx context.Context, query string) error {
-	if c == nil {
-		return nil
-	}
-	cfg := c.effectiveIcebergConfig()
-	if !shouldLoadIcebergColumnMetadata(cfg, c.physicalCatalog, c.passthrough) {
-		return nil
-	}
-	return icebergmeta.LoadColumns(ctx, c.executor, query, icebergmeta.Config{
-		LakekeeperEndpoint:        cfg.LakekeeperEndpoint,
-		LakekeeperWarehouse:       cfg.LakekeeperWarehouse,
-		LakekeeperOAuth2ServerURI: cfg.LakekeeperOAuth2ServerURI,
-	})
-}
-
-func (c *clientConn) effectiveIcebergConfig() IcebergConfig {
-	if c != nil && c.hasTenantIcebergConfig {
-		return c.tenantIcebergConfig
-	}
-	if c != nil && c.server != nil {
-		return c.server.cfg.Iceberg
-	}
-	return IcebergConfig{}
-}
-
-// shouldLoadIcebergColumnMetadata reports whether a query should trigger the
-// on-demand Iceberg column-metadata load. physicalCatalog is the session's
-// resolved DuckDB catalog (set during session init, and via
-// SetConnectionPhysicalCatalog on the control plane). The metadata load issues
-// Lakekeeper REST calls and DELETE/INSERT against the shared
-// memory.main.__duckgres_iceberg_column_metadata table, so it must only run for
-// Iceberg sessions — otherwise a DuckLake session on a dual-catalog worker would
-// trigger cross-catalog REST I/O and churn shared state it can never read (the
-// compat views gate reads on current_database()='iceberg').
-func shouldLoadIcebergColumnMetadata(cfg IcebergConfig, physicalCatalog string, passthrough bool) bool {
-	return !passthrough &&
-		physicalCatalog == iceberg.CatalogName &&
-		cfg.Enabled &&
-		cfg.ResolvedBackend() == iceberg.BackendLakekeeper &&
-		cfg.LakekeeperOAuth2ServerURI == ""
-}
-
-func (c *clientConn) queryWithMetadata(ctx context.Context, query string, run func() (RowSet, error)) (RowSet, error) {
-	if err := c.loadIcebergColumnMetadata(ctx, query); err != nil {
-		return nil, err
-	}
-	return run()
-}
-
-func (c *clientConn) queryContextWithMetadata(ctx context.Context, query string) (RowSet, error) {
-	return c.queryWithMetadata(ctx, query, func() (RowSet, error) {
-		return c.executor.QueryContext(ctx, query)
-	})
-}
-
-func (c *clientConn) queryWithArgsWithMetadata(ctx context.Context, query string, args ...interface{}) (RowSet, error) {
-	return c.queryWithMetadata(ctx, query, func() (RowSet, error) {
-		return c.executor.Query(query, args...)
-	})
-}
-
-// physicalDuckLakeCatalog is the physical catalog name DuckLake is attached as.
-const physicalDuckLakeCatalog = "ducklake"
-
-// executeSelectQuery runs a result-returning query against DuckDB and streams results to the client.
-// Sends RowDescription, DataRow messages, CommandComplete, and ReadyForQuery.
-// Returns the number of rows sent, any SQLSTATE+message sent to the client,
-// and any connection-level error.
-func (c *clientConn) executeSelectQuery(query string, cmdType string) (int64, string, string, error) {
-	ctx, cleanup := c.queryContext()
-	defer cleanup()
-
-	execStart := time.Now()
-	execCtx, execSpan := observe.Tracer().Start(ctx, "duckgres.execute")
-	// Lifecycle log pair: deferred logQueryFinished captures the eventual
-	// rowCount and any error from any return path — including Scan,
-	// ColumnTypes, sendRowDescription, and rows.Err() — so the pair is
-	// always balanced.
-	var queryRowsAff int64
-	var queryFinalErr error
-	c.logQueryStarted(query)
-	defer func() {
-		c.logQueryFinished(query, execStart, queryRowsAff, queryFinalErr)
-	}()
-	runQuery := func() (RowSet, error) {
-		return c.queryContextWithMetadata(ctx, query)
-	}
-
-	rows, err := runQuery()
-	if err != nil && c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-		ducklakeConflictTotal.Inc()
-		rows, err = retryOnConflict(runQuery)
-	}
-	if err != nil {
-		rows, err, _ = recoverAbortedTransaction(
-			err,
-			c.txStatus == txStatusIdle,
-			func() error {
-				_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-				return rollbackErr
-			},
-			func() (RowSet, error) {
-				return c.executor.QueryContext(ctx, query)
-			},
-		)
-	}
-	c.lastProfilingSummary = observe.EnrichSpanWithProfiling(execCtx, execSpan, execStart, c.executor, c.orgID)
-	execSpan.End()
-	if err != nil {
-		queryFinalErr = err
-		errCode := classifyErrorCode(err)
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errMsg = "canceling statement due to user request"
-		} else {
-			c.logQueryError(query, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return 0, errCode, errMsg, nil
-	}
-	defer func() { _ = rows.Close() }()
-
-	cols, err := rows.Columns()
-	if err != nil {
-		queryFinalErr = err
-		errCode := "42000"
-		errMsg := friendlyExecError(err)
-		if !c.isCallerCancellation(err) {
-			c.logQueryError(query, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return 0, errCode, errMsg, nil
-	}
-
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		queryFinalErr = err
-		errCode := "42000"
-		errMsg := friendlyExecError(err)
-		if !c.isCallerCancellation(err) {
-			c.logQueryError(query, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return 0, errCode, errMsg, nil
-	}
-
-	_, sendSpan := observe.Tracer().Start(ctx, "duckgres.send_results")
-	defer sendSpan.End()
-
-	// Mid-stream wire-write failures (sendRowDescription / sendDataRowWithFormats)
-	// surface a broken socket — typically the AWS NLB tearing down a stalled
-	// connection, or the kernel collapsing the socket after TCP_USER_TIMEOUT.
-	// They must route through logQueryError so the SQLSTATE-class severity
-	// router fires Error-level "Query execution errored." for alerts.
-	// Pre-fix the only signal was Info-level "Query finished." from the
-	// deferred lifecycle log, which silently disappears below alerting.
-	if err := c.sendRowDescription(cols, colTypes); err != nil {
-		queryFinalErr = err
-		if !c.isCallerCancellation(err) {
-			c.logQueryError(query, fmt.Errorf("pgwire client write failed sending row description: %w", err))
-		}
-		return 0, "", "", err
-	}
-
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	rowCount := 0
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			queryFinalErr = err
-			errCode := "42000"
-			errMsg := friendlyExecError(err)
-			if !c.isCallerCancellation(err) {
-				c.logQueryError(query, err)
-			}
-			c.sendError("ERROR", errCode, errMsg)
-			c.setTxError()
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return 0, errCode, errMsg, nil
-		}
-
-		if err := c.sendDataRowWithFormats(values, nil, typeOIDs); err != nil {
-			queryFinalErr = err
-			if !c.isCallerCancellation(err) {
-				c.logQueryError(query, fmt.Errorf("pgwire client write failed during result streaming: %w", err))
-			}
-			return 0, "", "", err
-		}
-		rowCount++
-	}
-	queryRowsAff = int64(rowCount)
-
-	if err := rows.Err(); err != nil {
-		queryFinalErr = err
-		errCode := "42000"
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errCode = "57014"
-			errMsg = "canceling statement due to user request"
-		} else {
-			c.logQueryError(query, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return 0, errCode, errMsg, nil
-	}
-
-	c.updateTxStatus(cmdType)
-	tag := buildCommandTagFromRowCount(cmdType, int64(rowCount))
-	_ = wire.WriteCommandComplete(c.writer, tag)
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return int64(rowCount), "", "", nil
-}
-
-// handleMultiStatementQuery processes multiple semicolon-separated statements
-// from a single Q (simple query) message. Per the PostgreSQL wire protocol,
-// each statement gets its own RowDescription/DataRow/CommandComplete messages,
-// with a single ReadyForQuery at the end. If any statement fails, remaining
-// statements are skipped.
-func (c *clientConn) handleMultiStatementQuery(query string) error {
-	// Re-parse with long identifiers protected so that splitting the batch via
-	// Deparse below does not silently truncate names > 63 bytes (see
-	// transpiler/longident.go). executeSingleStatement transpiles each
-	// statement again, which re-applies the same protection.
-	parseSQL, longIdents := transpiler.ProtectLongIdentifiers(query)
-	tree, err := pg_query.Parse(parseSQL)
-	if err != nil {
-		c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	slog.Debug("Multi-statement simple query.", "user", c.username, "count", len(tree.Stmts))
-
-	for _, stmt := range tree.Stmts {
-		// Deparse individual statement back to SQL
-		singleTree := &pg_query.ParseResult{
-			Stmts: []*pg_query.RawStmt{stmt},
-		}
-		singleSQL, err := pg_query.Deparse(singleTree)
-		if err != nil {
-			c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-			break
-		}
-		singleSQL = transpiler.RestoreLongIdentifiers(singleSQL, longIdents)
-
-		errSent, fatalErr := c.executeSingleStatement(singleSQL)
-		if fatalErr != nil {
-			return fatalErr
-		}
-		if errSent {
-			break // Stop processing remaining statements on error
-		}
-	}
-
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// executeSingleStatement transpiles and executes a single SQL statement,
-// sending results to the client. Does NOT send ReadyForQuery (the caller
-// is responsible for that). Returns (true, nil) if an error was sent to the
-// client (so the caller can stop processing a batch), or (false, err) for
-// fatal connection errors.
-func (c *clientConn) executeSingleStatement(query string) (errSent bool, fatalErr error) {
-	start := time.Now()
-
-	// Check for cursor operations before transpilation
-	tree, parseErr := pg_query.Parse(query)
-	if parseErr == nil && len(tree.Stmts) == 1 {
-		switch s := tree.Stmts[0].Stmt.Node.(type) {
-		case *pg_query.Node_DeclareCursorStmt:
-			innerSQL := deparseInnerQuery(s.DeclareCursorStmt.Query)
-			if innerSQL == "" {
-				c.sendError("ERROR", "42601", "could not deparse cursor query")
-				return true, nil
-			}
-			transpiledSQL := innerSQL
-			if !c.passthrough {
-				tr := c.newTranspiler(false)
-				result, err := tr.Transpile(innerSQL)
-				if err != nil {
-					c.sendError("ERROR", "42601", fmt.Sprintf("syntax error in cursor query: %v", err))
-					return true, nil
-				}
-				transpiledSQL = result.SQL
-				if result.FallbackToNative {
-					transpiledSQL = innerSQL
-				}
-			}
-			c.closeCursor(s.DeclareCursorStmt.Portalname)
-			c.cursors[s.DeclareCursorStmt.Portalname] = &cursorState{query: transpiledSQL}
-			_ = wire.WriteCommandComplete(c.writer, "DECLARE CURSOR")
-			return false, nil
-
-		case *pg_query.Node_FetchStmt:
-			if !isFetchForwardOnly(s.FetchStmt.Direction) {
-				c.sendError("ERROR", "0A000", "cursor can only scan forward")
-				return true, nil
-			}
-			cursor, ok := c.cursors[s.FetchStmt.Portalname]
-			if !ok {
-				c.sendError("ERROR", "34000", fmt.Sprintf("cursor %q does not exist", s.FetchStmt.Portalname))
-				return true, nil
-			}
-			if cursor.rows == nil {
-				if err := c.openCursor(cursor); err != nil {
-					if c.isCallerCancellation(err) {
-						c.sendError("ERROR", "57014", "canceling statement due to user request")
-					} else {
-						c.sendError("ERROR", "42000", err.Error())
-					}
-					c.setTxError()
-					return true, nil
-				}
-			}
-			howMany := s.FetchStmt.HowMany
-			if howMany < 0 {
-				c.sendError("ERROR", "0A000", "cursor can only scan forward")
-				return true, nil
-			}
-			if s.FetchStmt.Ismove {
-				moveCount := int64(0)
-				for moveCount < howMany && cursor.rows.Next() {
-					values := make([]interface{}, len(cursor.cols))
-					valuePtrs := make([]interface{}, len(cursor.cols))
-					for i := range values {
-						valuePtrs[i] = &values[i]
-					}
-					_ = cursor.rows.Scan(valuePtrs...)
-					moveCount++
-				}
-				_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("MOVE %d", moveCount))
-				return false, nil
-			}
-			if err := c.sendRowDescription(cursor.cols, cursor.colTypes); err != nil {
-				return false, err
-			}
-			rowCount := int64(0)
-			for rowCount < howMany && cursor.rows.Next() {
-				values := make([]interface{}, len(cursor.cols))
-				valuePtrs := make([]interface{}, len(cursor.cols))
-				for i := range values {
-					valuePtrs[i] = &values[i]
-				}
-				if err := cursor.rows.Scan(valuePtrs...); err != nil {
-					c.sendError("ERROR", "42000", err.Error())
-					c.setTxError()
-					return true, nil
-				}
-				if err := c.sendDataRowWithFormats(values, nil, cursor.typeOIDs); err != nil {
-					return false, err
-				}
-				rowCount++
-			}
-			if err := cursor.rows.Err(); err != nil {
-				if c.isCallerCancellation(err) {
-					c.sendError("ERROR", "57014", "canceling statement due to user request")
-				} else {
-					c.sendError("ERROR", "42000", err.Error())
-				}
-				c.setTxError()
-				return true, nil
-			}
-			_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("FETCH %d", rowCount))
-			return false, nil
-
-		case *pg_query.Node_ClosePortalStmt:
-			if s.ClosePortalStmt.Portalname == "" {
-				c.closeAllCursors()
-			} else {
-				if _, ok := c.cursors[s.ClosePortalStmt.Portalname]; !ok {
-					c.sendError("ERROR", "34000", fmt.Sprintf("cursor %q does not exist", s.ClosePortalStmt.Portalname))
-					return true, nil
-				}
-				c.closeCursor(s.ClosePortalStmt.Portalname)
-			}
-			_ = wire.WriteCommandComplete(c.writer, "CLOSE CURSOR")
-			return false, nil
-		}
-	}
-
-	// Intercept pg_cursors queries
-	if cursorName, _, ok := matchPgCursorsQuery(query); ok {
-		_, exists := c.cursors[cursorName]
-		_ = c.sendPgCursorsRowDescriptionWithFormats(nil)
-		rowCount := 0
-		if exists {
-			_ = c.sendDataRowWithFormats([]interface{}{int64(1)}, nil, []int32{23})
-			rowCount = 1
-		}
-		_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", rowCount))
-		return false, nil
-	}
-
-	// Intercept pg_stat_activity queries
-	if matchPgStatActivityQuery(query) {
-		_ = c.sendPgStatActivityRowDescriptionWithFormats(nil)
-		conns := c.server.listConns()
-		sort.Slice(conns, func(i, j int) bool { return conns[i].pid < conns[j].pid })
-		for _, conn := range conns {
-			_ = c.sendPgStatActivityDataRow(conn, nil)
-		}
-		_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", len(conns)))
-		return false, nil
-	}
-
-	// Transpile
-	tr := c.newTranspiler(false)
-	result, err := tr.Transpile(query)
-	if err != nil {
-		c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-		return true, nil
-	}
-
-	if result.FallbackToNative {
-		if err := c.validateWithDuckDB(query); err != nil {
-			c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-			return true, nil
-		}
-		slog.Debug("Fallback to native DuckDB.", "user", c.username, "query", query)
-	}
-
-	if result.Error != nil {
-		c.sendError("ERROR", transformErrorSQLState(result.Error), result.Error.Error())
-		return true, nil
-	}
-
-	if result.IsIgnoredSet {
-		_ = wire.WriteCommandComplete(c.writer, "SET")
-		return false, nil
-	}
-
-	if result.IsNoOp {
-		_ = wire.WriteCommandComplete(c.writer, result.NoOpTag)
-		return false, nil
-	}
-
-	// Multi-statement rewrites (writable CTEs) not supported inside batches
-	if len(result.Statements) > 0 {
-		c.sendError("ERROR", "0A000", "writable CTEs not supported in multi-statement queries")
-		return true, nil
-	}
-
-	executedQuery := c.rewriteDirectQuery(result.SQL)
-	if executedQuery != query {
-		slog.Debug("Query transpiled.", "user", c.username, "executed", executedQuery)
-	}
-
-	upperQuery := strings.ToUpper(executedQuery)
-	cmdType := c.getCommandType(upperQuery)
-
-	// COPY not supported inside batches
-	if cmdType == "COPY" {
-		c.sendError("ERROR", "0A000", "COPY not supported in multi-statement queries")
-		return true, nil
-	}
-
-	// Lifecycle log pair: every per-statement run in a multi-statement
-	// simple-query batch gets a logQueryStarted / logQueryFinished bracket
-	// (PR #519). The deferred close captures whichever code path the
-	// statement took — DML, SELECT, retry, transaction-conflict recovery.
-	queryStart := time.Now()
-	var queryRowsAff int64
-	var queryFinalErr error
-	c.logQueryStarted(executedQuery)
-	defer func() {
-		c.logQueryFinished(executedQuery, queryStart, queryRowsAff, queryFinalErr)
-	}()
-
-	if !queryReturnsResults(executedQuery) {
-		if cmdType == "BEGIN" && c.txStatus == txStatusTransaction {
-			c.sendNotice("WARNING", "25001", "there is already a transaction in progress")
-			_ = wire.WriteCommandComplete(c.writer, "BEGIN")
-			return false, nil
-		}
-
-		ctx, cleanup := c.queryContext()
-		defer cleanup()
-
-		runExec := func() (ExecResult, error) {
-			execResult, err := c.executor.ExecContext(ctx, executedQuery)
-			if err != nil {
-				fallbackResult, handled, fallbackErr := c.execCompatibilityFallback(ctx, executedQuery, err, func(fallbackQuery string) (ExecResult, error) {
-					return c.executor.ExecContext(ctx, fallbackQuery)
-				})
-				if handled {
-					return fallbackResult, fallbackErr
-				}
-			}
-			return execResult, err
-		}
-
-		execResult, err := runExec()
-		if err != nil {
-			if c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-				ducklakeConflictTotal.Inc()
-				execResult, err = retryOnConflict(runExec)
-			}
-			if err != nil {
-				execResult, err, _ = recoverAbortedTransaction(
-					err,
-					c.txStatus == txStatusIdle,
-					func() error {
-						_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-						return rollbackErr
-					},
-					runExec,
-				)
-			}
-			if err != nil {
-				queryFinalErr = err
-				errCode := classifyErrorCode(err)
-				errMsg := friendlyExecError(err)
-				if c.isCallerCancellation(err) {
-					errMsg = "canceling statement due to user request"
-				} else {
-					c.logQueryError(executedQuery, err)
-				}
-				c.sendError("ERROR", errCode, errMsg)
-				c.setTxError()
-				c.logQuery(start, query, executedQuery, cmdType, 0, 0, errCode, errMsg, "simple-batch")
-				return true, nil
-			}
-		}
-
-		var writtenRows int64
-		if execResult != nil {
-			writtenRows, _ = execResult.RowsAffected()
-		}
-		queryRowsAff = writtenRows
-		c.updateTxStatus(cmdType)
-		tag := c.buildCommandTag(cmdType, execResult)
-		_ = wire.WriteCommandComplete(c.writer, tag)
-		c.logQuery(start, query, executedQuery, cmdType, 0, writtenRows, "", "", "simple-batch")
-		return false, nil
-	}
-
-	// SELECT
-	ctx, cleanup := c.queryContext()
-	defer cleanup()
-
-	runQuery := func() (RowSet, error) {
-		return c.queryContextWithMetadata(ctx, executedQuery)
-	}
-
-	rows, err := runQuery()
-	if err != nil && c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-		ducklakeConflictTotal.Inc()
-		rows, err = retryOnConflict(runQuery)
-	}
-	if err != nil {
-		rows, err, _ = recoverAbortedTransaction(
-			err,
-			c.txStatus == txStatusIdle,
-			func() error {
-				_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-				return rollbackErr
-			},
-			runQuery,
-		)
-	}
-	if err != nil {
-		queryFinalErr = err
-		errCode := classifyErrorCode(err)
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errMsg = "canceling statement due to user request"
-		} else {
-			c.logQueryError(executedQuery, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		c.logQuery(start, query, executedQuery, cmdType, 0, 0, errCode, errMsg, "simple-batch")
-		return true, nil
-	}
-	defer func() { _ = rows.Close() }()
-
-	cols, err := rows.Columns()
-	if err != nil {
-		queryFinalErr = err
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		return true, nil
-	}
-
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		queryFinalErr = err
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		return true, nil
-	}
-
-	if err := c.sendRowDescription(cols, colTypes); err != nil {
-		queryFinalErr = err
-		return false, err
-	}
-
-	// Extract type OIDs for JSON-aware text formatting
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	rowCount := 0
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			queryFinalErr = err
-			c.sendError("ERROR", "42000", err.Error())
-			return true, nil
-		}
-
-		if err := c.sendDataRowWithFormats(values, nil, typeOIDs); err != nil {
-			queryFinalErr = err
-			return false, err
-		}
-		rowCount++
-	}
-	queryRowsAff = int64(rowCount)
-
-	c.updateTxStatus(cmdType)
-	tag := buildCommandTagFromRowCount(cmdType, int64(rowCount))
-	_ = wire.WriteCommandComplete(c.writer, tag)
-	c.logQuery(start, query, executedQuery, cmdType, int64(rowCount), 0, "", "", "simple-batch")
-	return false, nil
-}
-
-// executeMultiStatement handles execution of multi-statement query rewrites.
-// This is used for writable CTE transformations where a single PostgreSQL query
-// is rewritten into multiple DuckDB statements.
-//
-// Execution order:
-// 1. Execute setup statements (BEGIN, CREATE TEMP TABLE, etc.) - all but the last
-// 2. Execute final statement and obtain cursor (rows object)
-// 3. Execute cleanup statements (DROP temp tables, COMMIT) - cursor still valid
-// 4. Stream rows from cursor to client
-func (c *clientConn) executeMultiStatement(statements []string, cleanup []string) error {
-	if len(statements) == 0 {
-		_ = wire.WriteEmptyQueryResponse(c.writer)
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Check if we're adding our own transaction wrapper
-	hasOurTransaction := len(statements) >= 2 &&
-		strings.ToUpper(strings.TrimSpace(statements[0])) == "BEGIN" &&
-		len(cleanup) > 0 &&
-		strings.ToUpper(strings.TrimSpace(cleanup[len(cleanup)-1])) == "COMMIT"
-
-	// If already in a transaction, skip our BEGIN/COMMIT wrapper
-	if hasOurTransaction && c.txStatus == txStatusTransaction {
-		statements = statements[1:]        // Strip BEGIN
-		cleanup = cleanup[:len(cleanup)-1] // Strip COMMIT from cleanup
-	}
-
-	// Execute setup statements (all but last). Each step is its own
-	// logical query on the worker, so each gets its own logQueryStarted /
-	// logQueryFinished pair (PR #519).
-	for i := 0; i < len(statements)-1; i++ {
-		stmt := statements[i]
-		slog.Debug("Multi-stmt setup.", "user", c.username, "step", i+1, "total", len(statements)-1, "stmt", stmt)
-		setupStart := time.Now()
-		c.logQueryStarted(stmt)
-		result, err := c.executor.Exec(stmt)
-		var setupRows int64
-		if result != nil {
-			setupRows, _ = result.RowsAffected()
-		}
-		c.logQueryFinished(stmt, setupStart, setupRows, err)
-		if err != nil {
-			slog.Error("Multi-stmt setup error.", "user", c.username, "query", stmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			// On error, still try to cleanup (best effort)
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-	}
-
-	// Handle final statement
-	finalStmt := statements[len(statements)-1]
-	upperFinal := strings.ToUpper(strings.TrimSpace(finalStmt))
-	cmdType := c.getCommandType(upperFinal)
-	slog.Debug("Multi-stmt final.", "user", c.username, "stmt", finalStmt, "cmd_type", cmdType)
-
-	finalStart := time.Now()
-	var finalRowsAff int64
-	var finalErr error
-	c.logQueryStarted(finalStmt)
-	defer func() {
-		c.logQueryFinished(finalStmt, finalStart, finalRowsAff, finalErr)
-	}()
-
-	if queryReturnsResults(finalStmt) {
-		// Result-returning query: obtain cursor FIRST, cleanup SECOND, stream THIRD
-		rows, err := c.executor.Query(finalStmt)
-		if err != nil {
-			finalErr = err
-			slog.Error("Multi-stmt final query error.", "user", c.username, "query", finalStmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-		defer func() { _ = rows.Close() }()
-
-		// Execute cleanup while cursor is open (data is materialized in cursor)
-		// DuckDB cursor holds result data even after source tables are dropped
-		c.executeCleanup(cleanup)
-
-		// Now stream results from cursor. streamRowsToClient counts rows
-		// internally; we approximate Finished's rowsAff with 0 here
-		// (logQuery's own structured channel still records the precise
-		// count). A future refactor can plumb the count out of
-		// streamRowsToClient.
-		err = c.streamRowsToClient(rows, cmdType, finalStmt)
-		if err != nil {
-			finalErr = err
-		}
-		return err
-
-	} else {
-		// Non-result query (DML without RETURNING, DDL, etc.): execute then cleanup
-		result, err := c.executor.Exec(finalStmt)
-		if err != nil {
-			finalErr = err
-			slog.Error("Multi-stmt final exec error.", "user", c.username, "query", finalStmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-		if result != nil {
-			finalRowsAff, _ = result.RowsAffected()
-		}
-
-		// Execute cleanup
-		c.executeCleanup(cleanup)
-
-		// Send completion
-		tag := c.buildCommandTag(cmdType, result)
-		_ = wire.WriteCommandComplete(c.writer, tag)
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		return c.writer.Flush()
-	}
-}
-
-// executeCleanup runs cleanup statements, ignoring errors (best effort).
-// This is used to clean up temp tables after a multi-statement query.
-func (c *clientConn) executeCleanup(cleanup []string) {
-	for _, stmt := range cleanup {
-		slog.Debug("Multi-stmt cleanup.", "user", c.username, "stmt", stmt)
-		_, err := c.executor.Exec(stmt)
-		if err != nil {
-			// Log but don't fail - cleanup is best effort
-			slog.Warn("Multi-stmt cleanup error (ignored).", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		}
-	}
-}
-
-// executeMultiStatementExtended handles execution of multi-statement query rewrites
-// for the extended query protocol (Parse/Bind/Execute).
-// Unlike executeMultiStatement, this does NOT send ReadyForQuery (that's done by Sync).
-func (c *clientConn) executeMultiStatementExtended(statements []string, cleanup []string, args []interface{}, resultFormats []int16, described bool) {
-	if len(statements) == 0 {
-		_ = wire.WriteEmptyQueryResponse(c.writer)
-		return
-	}
-
-	// Check if we're adding our own transaction wrapper
-	hasOurTransaction := len(statements) >= 2 &&
-		strings.ToUpper(strings.TrimSpace(statements[0])) == "BEGIN" &&
-		len(cleanup) > 0 &&
-		strings.ToUpper(strings.TrimSpace(cleanup[len(cleanup)-1])) == "COMMIT"
-
-	// If already in a transaction, skip our BEGIN/COMMIT wrapper
-	if hasOurTransaction && c.txStatus == txStatusTransaction {
-		statements = statements[1:]        // Strip BEGIN
-		cleanup = cleanup[:len(cleanup)-1] // Strip COMMIT from cleanup
-	}
-
-	// Execute setup statements (all but last). Each step gets its own
-	// logQueryStarted / logQueryFinished pair (PR #519) so multi-stmt
-	// setup work is observable per statement, not just per outer query.
-	for i := 0; i < len(statements)-1; i++ {
-		stmt := statements[i]
-		slog.Debug("Multi-stmt-ext setup.", "user", c.username, "step", i+1, "total", len(statements)-1, "stmt", stmt)
-		setupStart := time.Now()
-		c.logQueryStarted(stmt)
-		result, err := c.executor.Exec(stmt, args...)
-		var setupRows int64
-		if result != nil {
-			setupRows, _ = result.RowsAffected()
-		}
-		c.logQueryFinished(stmt, setupStart, setupRows, err)
-		if err != nil {
-			slog.Error("Multi-stmt-ext setup error.", "user", c.username, "query", stmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			// On error, still try to cleanup (best effort)
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			return
-		}
-	}
-
-	// Handle final statement
-	finalStmt := statements[len(statements)-1]
-	upperFinal := strings.ToUpper(strings.TrimSpace(finalStmt))
-	cmdType := c.getCommandType(upperFinal)
-	slog.Debug("Multi-stmt-ext final.", "user", c.username, "stmt", finalStmt, "cmd_type", cmdType)
-
-	finalStart := time.Now()
-	var finalRowsAff int64
-	var finalErr error
-	c.logQueryStarted(finalStmt)
-	defer func() {
-		c.logQueryFinished(finalStmt, finalStart, finalRowsAff, finalErr)
-	}()
-
-	if queryReturnsResults(finalStmt) {
-		// Result-returning query: obtain cursor FIRST, cleanup SECOND, stream THIRD
-		rows, err := c.executor.Query(finalStmt, args...)
-		if err != nil {
-			finalErr = err
-			slog.Error("Multi-stmt-ext final query error.", "user", c.username, "query", finalStmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			return
-		}
-		defer func() { _ = rows.Close() }()
-
-		// Execute cleanup while cursor is open (data is materialized in cursor)
-		c.executeCleanup(cleanup)
-
-		// Stream results from cursor (extended protocol version). Row count
-		// is tracked by streamRowsToClientExtended; the deferred Finished
-		// log uses 0 as an approximation (logQuery still records the
-		// precise count via the structured channel).
-		c.streamRowsToClientExtended(rows, cmdType, resultFormats, described, finalStmt)
-
-	} else {
-		// Non-result query (DML without RETURNING, DDL, etc.): execute then cleanup
-		result, err := c.executor.Exec(finalStmt, args...)
-		if err != nil {
-			finalErr = err
-			slog.Error("Multi-stmt-ext final exec error.", "user", c.username, "query", finalStmt, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.setTxError()
-			c.executeCleanup(cleanup)
-			c.sendError("ERROR", "42000", err.Error())
-			return
-		}
-		if result != nil {
-			finalRowsAff, _ = result.RowsAffected()
-		}
-
-		// Execute cleanup
-		c.executeCleanup(cleanup)
-
-		// Send completion (no ReadyForQuery - that's done by Sync)
-		tag := c.buildCommandTag(cmdType, result)
-		_ = wire.WriteCommandComplete(c.writer, tag)
-	}
-}
-
-// streamRowsToClientExtended sends result rows for the extended query protocol.
-// Unlike streamRowsToClient, this does NOT send ReadyForQuery, and supports
-// binary result formats and the described flag.
-func (c *clientConn) streamRowsToClientExtended(rows RowSet, cmdType string, resultFormats []int16, described bool, query string) {
-	// Get column info
-	cols, err := rows.Columns()
-	if err != nil {
-		slog.Error("Failed to get column info.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		return
-	}
-
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		slog.Error("Failed to get column types.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		return
-	}
-
-	// Get type OIDs for binary encoding
-	typeOIDs := make([]int32, len(cols))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Send RowDescription if Describe wasn't called before Execute
-	if !described && len(cols) > 0 {
-		if err := c.sendRowDescriptionWithFormats(cols, colTypes, resultFormats); err != nil {
-			return
-		}
-	}
-
-	// Stream DataRows with format codes
-	rowCount := 0
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			slog.Error("Failed to scan row.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			return
-		}
-
-		if err := c.sendDataRowWithFormats(values, resultFormats, typeOIDs); err != nil {
-			return
-		}
-		rowCount++
-	}
-
-	if err := rows.Err(); err != nil {
-		if c.isCallerCancellation(err) {
-			c.sendError("ERROR", "57014", "canceling statement due to user request")
-		} else {
-			slog.Error("Row iteration error.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.sendError("ERROR", "42000", err.Error())
-		}
-		c.setTxError()
-		return
-	}
-
-	// Send completion (no ReadyForQuery - that's done by Sync)
-	tag := buildCommandTagFromRowCount(cmdType, int64(rowCount))
-	_ = wire.WriteCommandComplete(c.writer, tag)
-}
-
-// streamRowsToClient sends result rows over the wire protocol.
-// The rows cursor must already be obtained before calling this function.
-func (c *clientConn) streamRowsToClient(rows RowSet, cmdType string, query string) error {
-	// Get column info
-	cols, err := rows.Columns()
-	if err != nil {
-		slog.Error("Failed to get column info.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		slog.Error("Failed to get column types.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Send row description
-	if err := c.sendRowDescription(cols, colTypes); err != nil {
-		return err
-	}
-
-	// Extract type OIDs for JSON-aware text formatting
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Stream DataRows
-	rowCount := 0
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			slog.Error("Failed to scan row.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-
-		if err := c.sendDataRowWithFormats(values, nil, typeOIDs); err != nil {
-			return err
-		}
-		rowCount++
-	}
-
-	if err := rows.Err(); err != nil {
-		if c.isCallerCancellation(err) {
-			c.sendError("ERROR", "57014", "canceling statement due to user request")
-		} else {
-			slog.Error("Row iteration error.", "user", c.username, "query", query, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.sendError("ERROR", "42000", err.Error())
-		}
-		c.setTxError()
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Send completion
-	tag := buildCommandTagFromRowCount(cmdType, int64(rowCount))
-	_ = wire.WriteCommandComplete(c.writer, tag)
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	return c.writer.Flush()
-}
-
 // countDollarParams counts $N-style parameter placeholders in a query string
 // using a manual byte scan. This avoids pg_query.Parse which may fail on
 // DuckDB-native SQL syntax.
@@ -3564,1894 +2062,7 @@ func (c *clientConn) buildCommandTag(cmdType string, result ExecResult) string {
 	}
 }
 
-func shouldHandleCopyBeforeTranspile(query string) bool {
-	trimmed := strings.TrimSpace(query)
-	return strings.HasPrefix(strings.ToUpper(trimmed), "COPY")
-}
-
-// Regular expressions for parsing COPY commands
-var (
-	copyToStdoutRegex   = regexp.MustCompile(`(?i)COPY\s+(.+?)\s+TO\s+STDOUT`)
-	copyFromStdinRegex  = regexp.MustCompile(`(?i)COPY\s+(\S+)\s*(?:\(([^)]+)\)\s*)?FROM\s+STDIN`)
-	copyBinaryRegex     = regexp.MustCompile(`(?i)\b(?:STDIN|STDOUT)\b(?:[^;]*\bFORMAT\s+(?:"?binary"?|BINARY)\b|(?:\s+WITH)?\s+BINARY\b)`)
-	copyWithCSVRegex    = regexp.MustCompile(`(?i)\bCSV\b`)
-	copyWithHeaderRegex = regexp.MustCompile(`(?i)\bHEADER\b`)
-	copyDelimiterRegex  = regexp.MustCompile(`(?i)\bDELIMITER\s+['"](.)['"]\b`)
-	copyNullRegex       = regexp.MustCompile(`(?i)\bNULL\s+'([^']*)'`)
-	copyQuoteRegex      = regexp.MustCompile(`(?i)\bQUOTE\s+['"](.)['"]\s*`)
-	copyEscapeRegex     = regexp.MustCompile(`(?i)\bESCAPE\s+['"](.)['"]\s*`)
-)
-
-// stripPublicSchema maps PostgreSQL's "public" schema to DuckDB's default schema
-// in COPY table names. Handles both quoted ("public"."table") and unquoted (public.table) forms.
-// This mirrors the transpiler's PublicSchemaTransform but operates on raw table name strings
-// since COPY commands bypass the SQL transpiler (pg_query can't parse COPY ... FORMAT BINARY).
-func stripPublicSchema(tableName string) string {
-	// Quoted form: "public"."tablename" → "tablename"
-	if strings.HasPrefix(tableName, `"public".`) {
-		return tableName[len(`"public".`):]
-	}
-	// Unquoted form: public.tablename → tablename
-	if strings.HasPrefix(strings.ToLower(tableName), "public.") {
-		return tableName[len("public."):]
-	}
-	return tableName
-}
-
-// CopyFromOptions contains parsed options from a COPY FROM STDIN command
-type CopyFromOptions struct {
-	TableName  string
-	ColumnList string // Empty string or "(col1, col2, ...)"
-	Delimiter  string
-	HasHeader  bool
-	NullString string
-	Quote      string // Quote character (default " for CSV)
-	Escape     string // Escape character (default same as Quote)
-	IsBinary   bool   // True if FORMAT binary
-}
-
-// ParseCopyFromOptions extracts options from a COPY FROM STDIN command
-func ParseCopyFromOptions(query string) (*CopyFromOptions, error) {
-	upperQuery := strings.ToUpper(query)
-
-	matches := copyFromStdinRegex.FindStringSubmatch(query)
-	if len(matches) < 2 {
-		return nil, fmt.Errorf("invalid COPY FROM STDIN syntax")
-	}
-
-	opts := &CopyFromOptions{
-		TableName:  stripPublicSchema(matches[1]),
-		Delimiter:  "\t",  // Default PostgreSQL text format delimiter
-		NullString: "\\N", // Default PostgreSQL null representation
-	}
-
-	// Extract column list if present
-	if len(matches) > 2 && matches[2] != "" {
-		opts.ColumnList = fmt.Sprintf("(%s)", matches[2])
-	}
-
-	// Detect binary format
-	if copyBinaryRegex.MatchString(upperQuery) {
-		opts.IsBinary = true
-		return opts, nil
-	}
-
-	// Parse delimiter
-	if m := copyDelimiterRegex.FindStringSubmatch(query); len(m) > 1 {
-		opts.Delimiter = m[1]
-	} else if copyWithCSVRegex.MatchString(upperQuery) {
-		opts.Delimiter = ","
-	}
-
-	// Parse header option (only valid with CSV)
-	opts.HasHeader = copyWithCSVRegex.MatchString(upperQuery) && copyWithHeaderRegex.MatchString(upperQuery)
-
-	// Parse NULL string option
-	if m := copyNullRegex.FindStringSubmatch(query); len(m) > 1 {
-		opts.NullString = m[1]
-	}
-
-	// Parse QUOTE option (default " for CSV)
-	if m := copyQuoteRegex.FindStringSubmatch(query); len(m) > 1 {
-		opts.Quote = m[1]
-	} else if copyWithCSVRegex.MatchString(upperQuery) {
-		opts.Quote = `"` // Default quote character for CSV
-	}
-
-	// Parse ESCAPE option (default same as QUOTE)
-	if m := copyEscapeRegex.FindStringSubmatch(query); len(m) > 1 {
-		opts.Escape = m[1]
-	}
-
-	return opts, nil
-}
-
-// CopyToOptions contains parsed options from a COPY TO STDOUT command
-type CopyToOptions struct {
-	Source    string // Table name or (SELECT query)
-	Delimiter string
-	HasHeader bool
-	IsQuery   bool // True if Source is a query in parentheses
-}
-
-// ParseCopyToOptions extracts options from a COPY TO STDOUT command
-func ParseCopyToOptions(query string) (*CopyToOptions, error) {
-	upperQuery := strings.ToUpper(query)
-
-	matches := copyToStdoutRegex.FindStringSubmatch(query)
-	if len(matches) < 2 {
-		return nil, fmt.Errorf("invalid COPY TO STDOUT syntax")
-	}
-
-	source := strings.TrimSpace(matches[1])
-	opts := &CopyToOptions{
-		Source:    source,
-		Delimiter: "\t", // Default PostgreSQL text format delimiter
-		IsQuery:   strings.HasPrefix(source, "(") && strings.HasSuffix(source, ")"),
-	}
-
-	// Parse delimiter
-	if m := copyDelimiterRegex.FindStringSubmatch(query); len(m) > 1 {
-		opts.Delimiter = m[1]
-	} else if copyWithCSVRegex.MatchString(upperQuery) {
-		opts.Delimiter = ","
-	}
-
-	// Parse header option (only valid with CSV)
-	opts.HasHeader = copyWithCSVRegex.MatchString(upperQuery) && copyWithHeaderRegex.MatchString(upperQuery)
-
-	return opts, nil
-}
-
-// BuildDuckDBCopyFromSQL generates a DuckDB COPY FROM statement
-func BuildDuckDBCopyFromSQL(tableName, columnList, filePath string, opts *CopyFromOptions) string {
-	// DuckDB syntax: COPY table FROM 'file' (FORMAT CSV, HEADER, NULL 'value', DELIMITER ',', QUOTE '"')
-	// AUTO_DETECT FALSE disables sniffer to prevent it from overriding our settings
-	// STRICT_MODE FALSE allows reading rows that don't strictly comply with CSV standard
-	// PARALLEL FALSE avoids "Parallel CSV Reader does not support full read" errors
-	// on files streamed from COPY FROM STDIN (temp files with no seek support for sniffing)
-	copyOptions := []string{"FORMAT CSV", "AUTO_DETECT FALSE", "STRICT_MODE FALSE", "PARALLEL FALSE", "MAX_LINE_SIZE 10485760"}
-	if opts.HasHeader {
-		copyOptions = append(copyOptions, "HEADER")
-	}
-	// Always specify NULL string - DuckDB doesn't recognize \N by default
-	copyOptions = append(copyOptions, fmt.Sprintf("NULL '%s'", opts.NullString))
-	// Always specify DELIMITER explicitly (required when AUTO_DETECT is FALSE)
-	copyOptions = append(copyOptions, fmt.Sprintf("DELIMITER '%s'", opts.Delimiter))
-	// Always specify QUOTE for CSV to ensure proper quote handling
-	if opts.Quote != "" {
-		copyOptions = append(copyOptions, fmt.Sprintf("QUOTE '%s'", opts.Quote))
-		// Set ESCAPE to match QUOTE for RFC 4180 compliance (doubled quotes = escaped quote)
-		escape := opts.Escape
-		if escape == "" {
-			escape = opts.Quote
-		}
-		copyOptions = append(copyOptions, fmt.Sprintf("ESCAPE '%s'", escape))
-	} else if opts.Escape != "" {
-		copyOptions = append(copyOptions, fmt.Sprintf("ESCAPE '%s'", opts.Escape))
-	}
-
-	return fmt.Sprintf("COPY %s %s FROM '%s' (%s)",
-		tableName, columnList, filePath, strings.Join(copyOptions, ", "))
-}
-
-// handleCopy handles COPY TO STDOUT and COPY FROM STDIN commands
-func (c *clientConn) handleCopy(query, upperQuery string) error {
-	start := time.Now()
-
-	// Check if it's COPY TO STDOUT
-	if copyToStdoutRegex.MatchString(upperQuery) {
-		return c.handleCopyOut(query, upperQuery)
-	}
-
-	// Check if it's COPY FROM STDIN
-	if copyFromStdinRegex.MatchString(upperQuery) {
-		return c.handleCopyIn(query, upperQuery)
-	}
-
-	// For other COPY commands (e.g., COPY TO file), pass through to DuckDB
-	c.logQueryStarted(query)
-	queryStart := time.Now()
-	result, err := c.executor.Exec(query)
-	var rowsAffected int64
-	if result != nil {
-		rowsAffected, _ = result.RowsAffected()
-	}
-	c.logQueryFinished(query, queryStart, rowsAffected, err)
-	if err != nil {
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowsAffected))
-	c.logQuery(start, query, query, "COPY", 0, rowsAffected, "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleCopyOut handles COPY ... TO STDOUT
-func (c *clientConn) handleCopyOut(query, upperQuery string) error {
-	start := time.Now()
-	matches := copyToStdoutRegex.FindStringSubmatch(query)
-	if len(matches) < 2 {
-		c.sendError("ERROR", "42601", "Invalid COPY TO STDOUT syntax")
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42601", "Invalid COPY TO STDOUT syntax", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// The source can be a table name or a query in parentheses
-	source := strings.TrimSpace(matches[1])
-	var selectQuery string
-	if strings.HasPrefix(source, "(") && strings.HasSuffix(source, ")") {
-		selectQuery = source[1 : len(source)-1]
-	} else {
-		selectQuery = fmt.Sprintf("SELECT * FROM %s", source)
-	}
-
-	// Transpile the inner SELECT to handle schema mappings (e.g., public -> main)
-	// The outer COPY statement may not have been transpiled if pg_query can't parse
-	// the full COPY syntax (e.g., FORMAT "binary").
-	// Skip for passthrough users who send DuckDB-native SQL.
-	if !c.passthrough {
-		tr := c.newTranspiler(false)
-		if result, err := tr.Transpile(selectQuery); err == nil && !result.FallbackToNative {
-			selectQuery = result.SQL
-		}
-	}
-
-	// Execute the query. Lifecycle log pair (PR #519): every COPY-OUT
-	// driven SELECT gets a logQueryStarted and a logQueryFinished — the
-	// row count comes from the iteration loop further down, so the
-	// deferred close in the outer function path captures it.
-	queryStart := time.Now()
-	var copyRowsRead int64
-	var copyFinalErr error
-	c.logQueryStarted(selectQuery)
-	defer func() {
-		c.logQueryFinished(selectQuery, queryStart, copyRowsRead, copyFinalErr)
-	}()
-	rows, err := c.executor.Query(selectQuery)
-	if err != nil {
-		copyFinalErr = err
-		slog.Error("COPY TO query failed.", "user", c.username, "query", selectQuery, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	defer func() { _ = rows.Close() }()
-
-	cols, err := rows.Columns()
-	if err != nil {
-		slog.Error("COPY TO failed to get columns.", "user", c.username, "query", selectQuery, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	isBinary := copyBinaryRegex.MatchString(query)
-
-	if isBinary {
-		return c.handleCopyOutBinary(query, rows, cols)
-	}
-
-	// Get column types for JSON-aware formatting
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		copyFinalErr = err
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Parse text/CSV options
-	delimiter := "\t"
-	if m := copyDelimiterRegex.FindStringSubmatch(query); len(m) > 1 {
-		delimiter = m[1]
-	} else if copyWithCSVRegex.MatchString(upperQuery) {
-		delimiter = ","
-	}
-
-	// Send CopyOutResponse (text format)
-	if err := wire.WriteCopyOutResponse(c.writer, int16(len(cols)), true); err != nil {
-		return err
-	}
-	_ = c.writer.Flush()
-
-	// Send header if CSV with HEADER
-	if copyWithCSVRegex.MatchString(upperQuery) && copyWithHeaderRegex.MatchString(upperQuery) {
-		header := strings.Join(cols, delimiter) + "\n"
-		if err := wire.WriteCopyData(c.writer, []byte(header)); err != nil {
-			return err
-		}
-	}
-
-	// Send data rows
-	rowCount := 0
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			copyFinalErr = err
-			c.sendError("ERROR", "42000", err.Error())
-			c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "42000", err.Error(), "simple")
-			break
-		}
-
-		// Format row as tab/comma separated values
-		var rowData []string
-		for i, v := range values {
-			if typeOIDs[i] == OidJSON || typeOIDs[i] == OidJSONB {
-				rowData = append(rowData, string(encodeJSON(v)))
-			} else {
-				rowData = append(rowData, c.formatCopyValue(v))
-			}
-		}
-		line := strings.Join(rowData, delimiter) + "\n"
-		if err := wire.WriteCopyData(c.writer, []byte(line)); err != nil {
-			copyFinalErr = err
-			return err
-		}
-		rowCount++
-	}
-	copyRowsRead = int64(rowCount)
-
-	if err := rows.Err(); err != nil {
-		copyFinalErr = err
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Send CopyDone
-	if err := wire.WriteCopyDone(c.writer); err != nil {
-		copyFinalErr = err
-		return err
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-	c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleCopyOutBinary handles COPY ... TO STDOUT (FORMAT binary)
-// Implements PostgreSQL's binary COPY format: header, binary-encoded tuples, trailer.
-// Sends one CopyData message per tuple, with header prepended to the first tuple
-// and trailer appended to the last, matching how clients like DuckDB's postgres
-// extension consume binary COPY streams via PQgetCopyData.
-func (c *clientConn) handleCopyOutBinary(query string, rows RowSet, cols []string) error {
-	start := time.Now()
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, 0, "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Get type OIDs for each column
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Send CopyOutResponse (binary format)
-	if err := wire.WriteCopyOutResponse(c.writer, int16(len(cols)), false); err != nil {
-		return err
-	}
-	_ = c.writer.Flush()
-
-	// Binary COPY header (19 bytes)
-	binaryHeader := []byte{
-		'P', 'G', 'C', 'O', 'P', 'Y', '\n', 0xFF, '\r', '\n', 0x00,
-		0x00, 0x00, 0x00, 0x00,
-		0x00, 0x00, 0x00, 0x00,
-	}
-
-	// encodeTuple encodes a single tuple in PostgreSQL binary COPY format
-	encodeTuple := func(values []interface{}) []byte {
-		var buf bytes.Buffer
-		_ = binary.Write(&buf, binary.BigEndian, int16(len(values)))
-		for i, v := range values {
-			if v == nil {
-				_ = binary.Write(&buf, binary.BigEndian, int32(-1))
-			} else {
-				data := encodeBinary(v, typeOIDs[i])
-				if data == nil {
-					_ = binary.Write(&buf, binary.BigEndian, int32(-1))
-				} else {
-					_ = binary.Write(&buf, binary.BigEndian, int32(len(data)))
-					buf.Write(data)
-				}
-			}
-		}
-		return buf.Bytes()
-	}
-
-	// Send each tuple as its own CopyData message.
-	// The header is prepended to the first tuple's message.
-	rowCount := 0
-	firstRow := true
-	for rows.Next() {
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "42000", err.Error(), "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-
-		tupleBytes := encodeTuple(values)
-
-		if firstRow {
-			// First CopyData message: header + tuple
-			msg := make([]byte, 0, int64(len(binaryHeader))+int64(len(tupleBytes)))
-			msg = append(msg, binaryHeader...)
-			msg = append(msg, tupleBytes...)
-			if err := wire.WriteCopyData(c.writer, msg); err != nil {
-				return err
-			}
-			firstRow = false
-		} else {
-			if err := wire.WriteCopyData(c.writer, tupleBytes); err != nil {
-				return err
-			}
-		}
-		rowCount++
-	}
-
-	if err := rows.Err(); err != nil {
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "42000", err.Error(), "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// If no rows, still need to send header + trailer
-	if firstRow {
-		// No rows at all: send header + trailer in one message
-		msg := make([]byte, 0, len(binaryHeader)+2)
-		msg = append(msg, binaryHeader...)
-		msg = append(msg, 0xFF, 0xFF) // trailer: -1 as int16
-		if err := wire.WriteCopyData(c.writer, msg); err != nil {
-			return err
-		}
-	} else {
-		// Send trailer as its own CopyData message
-		if err := wire.WriteCopyData(c.writer, []byte{0xFF, 0xFF}); err != nil {
-			return err
-		}
-	}
-
-	// Send CopyDone
-	if err := wire.WriteCopyDone(c.writer); err != nil {
-		return err
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-	c.logQuery(start, query, query, "COPY", 0, int64(rowCount), "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleCopyIn handles COPY ... FROM STDIN
-func (c *clientConn) handleCopyIn(query, upperQuery string) error {
-	copyStartTime := time.Now()
-	slog.Debug("COPY FROM STDIN starting.", "user", c.username, "query", query)
-
-	// Parse COPY options using the helper function
-	opts, err := ParseCopyFromOptions(query)
-	if err != nil {
-		c.sendError("ERROR", "42601", "Invalid COPY FROM STDIN syntax")
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "42601", "Invalid COPY FROM STDIN syntax", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	tableName := opts.TableName
-	columnList := opts.ColumnList
-	slog.Debug("COPY FROM STDIN parsed.", "user", c.username, "table", tableName, "columns", columnList, "binary", opts.IsBinary)
-
-	// Get column info. If a column list is specified, query only those columns
-	// in the specified order to match the binary data field order.
-	var colQuery string
-	if columnList != "" {
-		// columnList is "(col1, col2, ...)" — use it in SELECT to get types in COPY order
-		colQuery = fmt.Sprintf("SELECT %s FROM %s LIMIT 0", columnList[1:len(columnList)-1], tableName)
-	} else {
-		colQuery = fmt.Sprintf("SELECT * FROM %s LIMIT 0", tableName)
-	}
-	testRows, err := c.executor.Query(colQuery)
-	if err != nil {
-		slog.Error("COPY FROM table check failed.", "user", c.username, "table", tableName, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		errMsg := fmt.Sprintf("relation \"%s\" does not exist", tableName)
-		c.sendError("ERROR", "42P01", errMsg)
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "42P01", errMsg, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	cols, _ := testRows.Columns()
-	colTypes, _ := testRows.ColumnTypes()
-	_ = testRows.Close()
-
-	// Branch to binary handler if binary format
-	if opts.IsBinary {
-		return c.handleCopyInBinary(query, opts, cols, colTypes)
-	}
-
-	// Check for BLOB columns. DuckDB's CSV parser cannot handle raw binary data
-	// in BLOB columns — it auto-detects the type but fails to parse the bytes.
-	// Fall back to in-process CSV parsing with batched INSERT for these tables.
-	var blobColIndices []int
-	for i, ct := range colTypes {
-		if ct.DatabaseTypeName() == "BLOB" {
-			blobColIndices = append(blobColIndices, i)
-		}
-	}
-	if len(blobColIndices) > 0 {
-		slog.Debug("COPY FROM STDIN: table has BLOB columns, using CSV parse fallback.", "user", c.username, "blob_columns", len(blobColIndices))
-		return c.handleCopyInCSVWithBlob(query, opts, cols, colTypes, blobColIndices)
-	}
-
-	// Send CopyInResponse
-	if err := wire.WriteCopyInResponse(c.writer, int16(len(cols)), true); err != nil {
-		return err
-	}
-	_ = c.writer.Flush()
-	slog.Debug("COPY FROM STDIN sent CopyInResponse, waiting for data.", "user", c.username)
-
-	// Remote-worker (Flight) executors implement CopyFromStdinExecutor so the
-	// CSV bytes are streamed to the worker pod via DoPut and spooled to the
-	// worker's filesystem there. The legacy local-tempfile path below works
-	// only when CP and worker share a filesystem (standalone / process
-	// backend), so prefer the streaming path when it's available.
-	if streamer, ok := c.executor.(sqlcore.CopyFromStdinExecutor); ok {
-		return c.handleCopyInRemoteStreaming(query, opts, copyStartTime, streamer)
-	}
-
-	// Create temp file upfront and stream data directly to it (avoids memory buffering).
-	// This approach leverages DuckDB's highly optimized CSV parser which handles
-	// type conversions automatically and can load millions of rows in seconds.
-	tmpFile, err := os.CreateTemp("", "duckgres-copy-*.csv")
-	if err != nil {
-		slog.Error("COPY FROM STDIN failed to create temp file.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		errMsg := fmt.Sprintf("failed to create temp file: %v", err)
-		c.sendError("ERROR", "58000", errMsg)
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "58000", errMsg, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	tmpPath := tmpFile.Name()
-	defer func() { _ = os.Remove(tmpPath) }()
-
-	// Stream COPY data directly to temp file (no memory buffering)
-	rowCount := 0
-	copyDataMessages := 0
-	bytesWritten := int64(0)
-	dataReceiveStart := time.Now()
-
-	for {
-		msgType, body, err := wire.ReadMessage(c.reader)
-		if err != nil {
-			slog.Error("COPY FROM STDIN error reading message.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			_ = tmpFile.Close()
-			return err
-		}
-
-		switch msgType {
-		case wire.MsgCopyData:
-			// Skip the PostgreSQL text COPY end-of-data marker (\.\n).
-			// Some clients send this as a CopyData message before CopyDone.
-			if (len(body) == 3 && body[0] == '\\' && body[1] == '.' && body[2] == '\n') ||
-				(len(body) == 2 && body[0] == '\\' && body[1] == '.') {
-				continue
-			}
-			n, err := tmpFile.Write(body)
-			if err != nil {
-				slog.Error("COPY FROM STDIN failed to write to temp file.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-				_ = tmpFile.Close()
-				errMsg := fmt.Sprintf("failed to write to temp file: %v", err)
-				c.sendError("ERROR", "58000", errMsg)
-				c.setTxError()
-				c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "58000", errMsg, "simple")
-				_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-				_ = c.writer.Flush()
-				return nil
-			}
-			bytesWritten += int64(n)
-			copyDataMessages++
-			if copyDataMessages%10000 == 0 {
-				slog.Debug("COPY FROM STDIN progress.", "user", c.username, "messages", copyDataMessages, "bytes", bytesWritten)
-			}
-
-		case wire.MsgCopyDone:
-			_ = tmpFile.Close()
-			dataReceiveElapsed := time.Since(dataReceiveStart)
-			slog.Debug("COPY FROM STDIN CopyDone received.", "user", c.username, "messages", copyDataMessages, "bytes", bytesWritten, "duration", dataReceiveElapsed)
-
-			// Build DuckDB COPY FROM statement using the helper function
-			copySQL := BuildDuckDBCopyFromSQL(tableName, columnList, tmpPath, opts)
-
-			slog.Debug("COPY FROM STDIN executing native DuckDB COPY.", "user", c.username, "sql", copySQL)
-			loadStart := time.Now()
-
-			// Lifecycle log pair (PR #519): the native DuckDB COPY FROM is
-			// the actual query the worker runs; everything before this is
-			// CSV byte-pumping into a tempfile, not worker work.
-			c.logQueryStarted(copySQL)
-			result, err := c.executor.Exec(copySQL)
-			var copyRowsAffected int64
-			if result != nil {
-				copyRowsAffected, _ = result.RowsAffected()
-			}
-			c.logQueryFinished(copySQL, loadStart, copyRowsAffected, err)
-			if err != nil {
-				slog.Error("COPY FROM STDIN DuckDB COPY failed.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-				errMsg := fmt.Sprintf("COPY failed: %v", err)
-				c.sendError("ERROR", "22P02", errMsg)
-				c.setTxError()
-				c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "22P02", errMsg, "simple")
-				_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-				_ = c.writer.Flush()
-				return nil
-			}
-
-			rowCount = int(copyRowsAffected)
-
-			totalElapsed := time.Since(copyStartTime)
-			loadElapsed := time.Since(loadStart)
-			slog.Info("COPY FROM STDIN completed.", "user", c.username, "rows", rowCount, "total_duration", totalElapsed, "load_duration", loadElapsed)
-
-			_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-			c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "", "", "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		case wire.MsgCopyFail:
-			// Client cancelled COPY
-			errMsg := string(bytes.TrimRight(body, "\x00"))
-			exception := fmt.Sprintf("COPY failed: %s", errMsg)
-			c.sendError("ERROR", "57014", exception)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "57014", exception, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		default:
-			errMsg := fmt.Sprintf("unexpected message type during COPY: %c", msgType)
-			c.sendError("ERROR", "08P01", errMsg)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "08P01", errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-	}
-}
-
-// handleCopyInRemoteStreaming handles COPY FROM STDIN when the executor is
-// remote (e.g. Flight, multitenant K8s worker). The control plane does not
-// share a filesystem with the worker pod, so the legacy "spool to local
-// /tmp, run COPY FROM <path>" approach fails with "No files found". This
-// path streams the wire CopyData bytes through the executor's
-// CopyFromStdin method, which ships them to the worker via Flight DoPut
-// and runs the COPY against a worker-local spool file.
-func (c *clientConn) handleCopyInRemoteStreaming(
-	query string,
-	opts *CopyFromOptions,
-	copyStartTime time.Time,
-	streamer sqlcore.CopyFromStdinExecutor,
-) error {
-	tableName := opts.TableName
-	columnList := opts.ColumnList
-
-	// Build the COPY SQL with the path placeholder; the worker substitutes
-	// in its own tempfile path before executing.
-	copySQL := BuildDuckDBCopyFromSQL(tableName, columnList, flightclient.CopyFromStdinPathPlaceholder, opts)
-	slog.Debug("COPY FROM STDIN streaming to remote worker.",
-		"user", c.username, "sql", copySQL, "worker", c.workerID, "worker_pod", c.workerPod)
-
-	r := &copyDataWireReader{c: c}
-
-	loadStart := time.Now()
-	c.logQueryStarted(copySQL)
-	rowCount, err := streamer.CopyFromStdin(c.ctx, copySQL, r)
-	c.logQueryFinished(copySQL, loadStart, rowCount, err)
-
-	// On wire-level CopyFail / unexpected message, the reader returns a
-	// non-EOF error so the streamer aborts the gRPC stream (its deferred
-	// cancel() prevents the worker from running COPY on partial bytes).
-	// We then surface the precise cause via the sticky flags below before
-	// falling through to the generic transport-error branch.
-	if r.cancelled {
-		exception := fmt.Sprintf("COPY failed: %s", r.cancelMsg)
-		c.sendError("ERROR", "57014", exception)
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "57014", exception, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	if r.protoErr != "" {
-		c.sendError("ERROR", "08P01", r.protoErr)
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "08P01", r.protoErr, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-	if err != nil {
-		slog.Error("COPY FROM STDIN remote streaming failed.",
-			"user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		errMsg := fmt.Sprintf("COPY failed: %v", err)
-		c.sendError("ERROR", "22P02", errMsg)
-		c.setTxError()
-		c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "22P02", errMsg, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	totalElapsed := time.Since(copyStartTime)
-	loadElapsed := time.Since(loadStart)
-	slog.Info("COPY FROM STDIN completed (remote streaming).",
-		"user", c.username, "rows", rowCount, "bytes", r.bytesRead,
-		"total_duration", totalElapsed, "load_duration", loadElapsed,
-		"worker", c.workerID, "worker_pod", c.workerPod)
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-	c.logQuery(copyStartTime, query, query, "COPY", 0, rowCount, "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// errCopyAborted is returned from copyDataWireReader.Read when the client
-// signalled CopyFail or sent an unexpected wire message mid-stream. The
-// streamer propagates this; the deferred gRPC context cancel then fires,
-// which causes the worker's Recv to fail with Canceled and skip running
-// the partially-uploaded COPY. handleCopyInRemoteStreaming inspects the
-// reader's sticky flags to decide the client-facing error code (57014 for
-// cancellation vs 08P01 for protocol error).
-var errCopyAborted = errors.New("copy data stream aborted")
-
-// copyDataWireReader adapts the PostgreSQL wire CopyData stream into an
-// io.Reader so it can be fed straight into a remote-worker upload. Behavior:
-//   - CopyDone           → io.EOF (clean end-of-stream)
-//   - CopyFail           → errCopyAborted, with cancelled / cancelMsg sticky
-//   - unexpected message → errCopyAborted, with protoErr sticky
-//   - wire transport err → that error is returned verbatim
-//
-// Returning a non-EOF error on cancellation matters for correctness:
-// if the reader pretended cancellation was a clean EOF, the streamer would
-// CloseSend cleanly and the worker would happily run COPY on the partial
-// bytes already received.
-type copyDataWireReader struct {
-	c *clientConn
-
-	pending   []byte
-	bytesRead int64
-
-	done      bool
-	cancelled bool
-	cancelMsg string
-	protoErr  string
-}
-
-func (r *copyDataWireReader) Read(p []byte) (int, error) {
-	for len(r.pending) == 0 {
-		if r.done {
-			if r.cancelled || r.protoErr != "" {
-				return 0, errCopyAborted
-			}
-			return 0, io.EOF
-		}
-		msgType, body, err := wire.ReadMessage(r.c.reader)
-		if err != nil {
-			r.done = true
-			return 0, err
-		}
-		switch msgType {
-		case wire.MsgCopyData:
-			// Skip the PostgreSQL text COPY end-of-data marker (\.\n).
-			if (len(body) == 3 && body[0] == '\\' && body[1] == '.' && body[2] == '\n') ||
-				(len(body) == 2 && body[0] == '\\' && body[1] == '.') {
-				continue
-			}
-			r.pending = body
-		case wire.MsgCopyDone:
-			r.done = true
-			return 0, io.EOF
-		case wire.MsgCopyFail:
-			r.done = true
-			r.cancelled = true
-			r.cancelMsg = string(bytes.TrimRight(body, "\x00"))
-			return 0, errCopyAborted
-		default:
-			r.done = true
-			r.protoErr = fmt.Sprintf("unexpected message type during COPY: %c", msgType)
-			return 0, errCopyAborted
-		}
-	}
-	n := copy(p, r.pending)
-	r.pending = r.pending[n:]
-	r.bytesRead += int64(n)
-	return n, nil
-}
-
-// handleCopyInCSVWithBlob handles COPY FROM STDIN for tables that contain BLOB columns.
-// DuckDB's native CSV COPY cannot handle raw binary data in BLOB columns because it
-// auto-detects the type and fails to parse the bytes. This method parses the CSV in Go,
-// converts BLOB column values to []byte, and uses batched INSERT statements.
-func (c *clientConn) handleCopyInCSVWithBlob(query string, opts *CopyFromOptions, cols []string, colTypes []ColumnTyper, blobColIndices []int) error {
-	copyStartTime := time.Now()
-
-	// Build a set for O(1) BLOB column index lookup
-	isBlobCol := make(map[int]bool, len(blobColIndices))
-	for _, idx := range blobColIndices {
-		isBlobCol[idx] = true
-	}
-
-	// Send CopyInResponse (text format)
-	if err := wire.WriteCopyInResponse(c.writer, int16(len(cols)), true); err != nil {
-		return err
-	}
-	_ = c.writer.Flush()
-
-	// Buffer all CopyData messages into memory (we need to parse CSV, not stream to file)
-	var buf bytes.Buffer
-	for {
-		msgType, body, err := wire.ReadMessage(c.reader)
-		if err != nil {
-			return err
-		}
-
-		switch msgType {
-		case wire.MsgCopyData:
-			// Skip end-of-data marker
-			if (len(body) == 3 && body[0] == '\\' && body[1] == '.' && body[2] == '\n') ||
-				(len(body) == 2 && body[0] == '\\' && body[1] == '.') {
-				continue
-			}
-			buf.Write(body)
-
-		case wire.MsgCopyDone:
-			dataReceiveElapsed := time.Since(copyStartTime)
-			slog.Debug("COPY FROM STDIN (BLOB fallback) CopyDone received.", "user", c.username, "bytes", buf.Len(), "duration", dataReceiveElapsed)
-
-			// Parse CSV from the buffered data
-			csvReader := csv.NewReader(&buf)
-			csvReader.Comma = rune(opts.Delimiter[0])
-			csvReader.LazyQuotes = true
-
-			// Skip header row if present
-			if opts.HasHeader {
-				if _, err := csvReader.Read(); err != nil {
-					errMsg := fmt.Sprintf("COPY failed: error reading CSV header: %v", err)
-					c.sendError("ERROR", "22P02", errMsg)
-					c.setTxError()
-					c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "22P02", errMsg, "simple")
-					_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-					_ = c.writer.Flush()
-					return nil
-				}
-			}
-
-			// Parse all rows and convert BLOB columns to []byte
-			var rows [][]interface{}
-			for {
-				record, err := csvReader.Read()
-				if err != nil {
-					break // EOF or error — stop reading
-				}
-				if len(record) != len(cols) {
-					slog.Warn("COPY FROM STDIN (BLOB fallback) skipping row with wrong field count.", "expected", len(cols), "got", len(record))
-					continue
-				}
-				row := make([]interface{}, len(cols))
-				for j, field := range record {
-					if field == opts.NullString {
-						row[j] = nil
-					} else if isBlobCol[j] {
-						row[j] = []byte(field)
-					} else {
-						row[j] = field
-					}
-				}
-				rows = append(rows, row)
-			}
-
-			if len(rows) == 0 {
-				_ = wire.WriteCommandComplete(c.writer, "COPY 0")
-				c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "", "", "simple")
-				_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-				_ = c.writer.Flush()
-				return nil
-			}
-
-			loadStart := time.Now()
-			rowCount, err := c.batchInsertRows(opts.TableName, opts.ColumnList, cols, rows)
-			if err != nil {
-				slog.Error("COPY FROM STDIN (BLOB fallback) INSERT failed.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-				errMsg := fmt.Sprintf("COPY failed: %v", err)
-				c.sendError("ERROR", "22P02", errMsg)
-				c.setTxError()
-				c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "22P02", errMsg, "simple")
-				_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-				_ = c.writer.Flush()
-				return nil
-			}
-
-			totalElapsed := time.Since(copyStartTime)
-			loadElapsed := time.Since(loadStart)
-			slog.Info("COPY FROM STDIN (BLOB fallback) completed.", "user", c.username, "rows", rowCount, "total_duration", totalElapsed, "load_duration", loadElapsed)
-
-			_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-			c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "", "", "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		case wire.MsgCopyFail:
-			errMsg := string(bytes.TrimRight(body, "\x00"))
-			exception := fmt.Sprintf("COPY failed: %s", errMsg)
-			c.sendError("ERROR", "57014", exception)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "57014", exception, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		default:
-			errMsg := fmt.Sprintf("unexpected message type during COPY: %c", msgType)
-			c.sendError("ERROR", "08P01", errMsg)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "08P01", errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-	}
-}
-
-// handleCopyInBinary handles COPY ... FROM STDIN with binary format.
-// It parses the PostgreSQL binary COPY format, decodes each field, and INSERTs rows.
-func (c *clientConn) handleCopyInBinary(query string, opts *CopyFromOptions, cols []string, colTypes []ColumnTyper) error {
-	copyStartTime := time.Now()
-
-	// Get type OIDs for decoding
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Send CopyInResponse (binary format)
-	if err := wire.WriteCopyInResponse(c.writer, int16(len(cols)), false); err != nil {
-		return err
-	}
-	_ = c.writer.Flush()
-	slog.Debug("COPY FROM STDIN binary: sent CopyInResponse.", "user", c.username)
-
-	// Collect all CopyData messages into a buffer
-	var buf bytes.Buffer
-	for {
-		msgType, body, err := wire.ReadMessage(c.reader)
-		if err != nil {
-			slog.Error("COPY FROM STDIN binary: error reading message.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			return err
-		}
-
-		switch msgType {
-		case wire.MsgCopyData:
-			buf.Write(body)
-
-		case wire.MsgCopyDone:
-			// Parse binary data and insert rows
-			data := buf.Bytes()
-			rowCount, err := c.parseBinaryCopyAndInsert(data, opts.TableName, opts.ColumnList, cols, typeOIDs)
-			if err != nil {
-				slog.Error("COPY FROM STDIN binary: parse/insert failed.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-				errMsg := fmt.Sprintf("COPY failed: %v", err)
-				c.sendError("ERROR", "22P02", errMsg)
-				c.setTxError()
-				c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "22P02", errMsg, "simple")
-				_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-				_ = c.writer.Flush()
-				return nil
-			}
-
-			elapsed := time.Since(copyStartTime)
-			slog.Info("COPY FROM STDIN binary completed.", "user", c.username, "rows", rowCount, "bytes", buf.Len(), "duration", elapsed)
-
-			_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("COPY %d", rowCount))
-			c.logQuery(copyStartTime, query, query, "COPY", 0, int64(rowCount), "", "", "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		case wire.MsgCopyFail:
-			errMsg := string(bytes.TrimRight(body, "\x00"))
-			exception := fmt.Sprintf("COPY failed: %s", errMsg)
-			c.sendError("ERROR", "57014", exception)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "57014", exception, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-
-		default:
-			errMsg := fmt.Sprintf("unexpected message type during COPY: %c", msgType)
-			c.sendError("ERROR", "08P01", errMsg)
-			c.setTxError()
-			c.logQuery(copyStartTime, query, query, "COPY", 0, 0, "08P01", errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-	}
-}
-
-// splitQualifiedName splits a possibly-quoted SQL name like "schema"."table" into parts.
-func splitQualifiedName(name string) []string {
-	var parts []string
-	var current strings.Builder
-	inQuotes := false
-	for i := 0; i < len(name); i++ {
-		if name[i] == '"' {
-			inQuotes = !inQuotes
-			continue
-		}
-		if name[i] == '.' && !inQuotes {
-			parts = append(parts, current.String())
-			current.Reset()
-			continue
-		}
-		current.WriteByte(name[i])
-	}
-	parts = append(parts, current.String())
-	return parts
-}
-
-// appendWithDuckDBAppender uses the DuckDB Appender API for fast bulk inserts.
-// Only works for full-column inserts (no column subset).
-// batchInsertRows inserts rows using batched multi-row INSERT statements.
-// Used as fallback when Appender can't be used (column subsets, unsupported types).
-func (c *clientConn) batchInsertRows(tableName, columnList string, cols []string, rows [][]interface{}) (int, error) {
-	const batchSize = 1000
-	numCols := len(cols)
-
-	colNames := columnList
-	if colNames == "" {
-		quotedCols := make([]string, numCols)
-		for i, col := range cols {
-			quotedCols[i] = fmt.Sprintf(`"%s"`, col)
-		}
-		colNames = "(" + strings.Join(quotedCols, ", ") + ")"
-	}
-
-	rowCount := 0
-	for start := 0; start < len(rows); start += batchSize {
-		end := start + batchSize
-		if end > len(rows) {
-			end = len(rows)
-		}
-		batch := rows[start:end]
-
-		var valueClauses []string
-		var args []interface{}
-		paramIdx := 1
-		for _, row := range batch {
-			placeholders := make([]string, numCols)
-			for j := range row {
-				placeholders[j] = fmt.Sprintf("$%d", paramIdx)
-				args = append(args, row[j])
-				paramIdx++
-			}
-			valueClauses = append(valueClauses, "("+strings.Join(placeholders, ", ")+")")
-		}
-
-		insertSQL := fmt.Sprintf("INSERT INTO %s %s VALUES %s",
-			tableName, colNames, strings.Join(valueClauses, ", "))
-
-		// Lifecycle log pair (PR #519) per batched INSERT — one query
-		// the worker actually runs.
-		batchStart := time.Now()
-		c.logQueryStarted(insertSQL)
-		result, err := c.executor.Exec(insertSQL, args...)
-		var rowsAff int64
-		if result != nil {
-			rowsAff, _ = result.RowsAffected()
-		}
-		c.logQueryFinished(insertSQL, batchStart, rowsAff, err)
-		if err != nil {
-			return rowCount, fmt.Errorf("batch INSERT failed at rows %d-%d: %v", start+1, start+len(batch), err)
-		}
-		rowCount += len(batch)
-	}
-
-	return rowCount, nil
-}
-
-// parseBinaryCopyAndInsert parses PostgreSQL binary COPY format data and inserts rows.
-// Uses the DuckDB Appender API for full-column inserts (fast path), falling back to
-// batched multi-row INSERT for column subsets or unsupported types.
-func (c *clientConn) parseBinaryCopyAndInsert(data []byte, tableName, columnList string, cols []string, typeOIDs []int32) (int, error) {
-	offset := 0
-
-	// Validate and skip header (19+ bytes)
-	// Signature: "PGCOPY\n\377\r\n\0" (11 bytes)
-	if len(data) < 19 {
-		return 0, fmt.Errorf("binary COPY data too short for header")
-	}
-	expectedSig := []byte{'P', 'G', 'C', 'O', 'P', 'Y', '\n', 0xFF, '\r', '\n', 0x00}
-	if !bytes.Equal(data[:11], expectedSig) {
-		return 0, fmt.Errorf("invalid binary COPY signature")
-	}
-	offset = 11
-
-	// Flags (4 bytes) and extension area length (4 bytes)
-	offset += 4
-	extLen := binary.BigEndian.Uint32(data[offset:])
-	offset += 4
-	offset += int(extLen) // skip extension area
-
-	numCols := len(cols)
-
-	// Parse all rows from binary data first
-	var rows [][]interface{}
-	for offset < len(data) {
-		if offset+2 > len(data) {
-			return len(rows), fmt.Errorf("truncated binary COPY data at tuple header")
-		}
-
-		fieldCount := int16(binary.BigEndian.Uint16(data[offset:]))
-		offset += 2
-
-		// Trailer: field count of -1
-		if fieldCount == -1 {
-			break
-		}
-
-		if int(fieldCount) != numCols {
-			return len(rows), fmt.Errorf("binary COPY field count mismatch: got %d, expected %d", fieldCount, numCols)
-		}
-
-		values := make([]interface{}, numCols)
-		for i := 0; i < numCols; i++ {
-			if offset+4 > len(data) {
-				return len(rows), fmt.Errorf("truncated binary COPY data at field %d length", i)
-			}
-
-			fieldLen := int32(binary.BigEndian.Uint32(data[offset:]))
-			offset += 4
-
-			if fieldLen == -1 {
-				values[i] = nil
-			} else {
-				if offset+int(fieldLen) > len(data) {
-					return len(rows), fmt.Errorf("truncated binary COPY data at field %d data", i)
-				}
-				fieldData := data[offset : offset+int(fieldLen)]
-				offset += int(fieldLen)
-
-				decoded, err := decodeBinaryCopy(fieldData, typeOIDs[i])
-				if err != nil {
-					return len(rows), fmt.Errorf("failed to decode field %d (OID %d, %d bytes): %v", i, typeOIDs[i], fieldLen, err)
-				}
-				values[i] = decoded
-			}
-		}
-		rows = append(rows, values)
-	}
-
-	if len(rows) == 0 {
-		return 0, nil
-	}
-
-	// Fast path: use Appender for full-column inserts (no column subset)
-	if columnList == "" {
-		count, err := c.appendWithDuckDBAppender(tableName, rows)
-		if err == nil {
-			return count, nil
-		}
-		slog.Warn("Appender failed, falling back to batched INSERT.", "table", tableName, "error", err)
-	}
-
-	// Fallback: batched multi-row INSERT
-	return c.batchInsertRows(tableName, columnList, cols, rows)
-}
-
-// decodeBinaryCopy decodes a binary COPY field, using field length to resolve type ambiguity.
-// DuckDB's postgres extension may send different integer widths than what the table OID suggests.
-func decodeBinaryCopy(data []byte, oid int32) (interface{}, error) {
-	if data == nil {
-		return nil, nil
-	}
-
-	// Zero-length fields: return empty string for text types, empty bytes for bytea, nil for others
-	if len(data) == 0 {
-		switch oid {
-		case OidText, OidVarchar, OidBpchar, OidName, OidJSON, OidJSONB:
-			return "", nil
-		case OidBytea:
-			return []byte{}, nil
-		default:
-			return nil, nil
-		}
-	}
-
-	switch oid {
-	case OidBool:
-		return decodeBool(data)
-	case OidInt2, OidInt4, OidInt8, OidOid:
-		// Use field length to determine actual integer width
-		switch len(data) {
-		case 2:
-			return decodeInt2(data)
-		case 4:
-			return decodeInt4(data)
-		case 8:
-			return decodeInt8(data)
-		default:
-			return string(data), nil
-		}
-	case OidFloat4:
-		if len(data) == 8 {
-			return decodeFloat8(data)
-		}
-		return decodeFloat4(data)
-	case OidFloat8:
-		if len(data) == 4 {
-			return decodeFloat4(data)
-		}
-		return decodeFloat8(data)
-	case OidNumeric:
-		return decodeNumeric(data)
-	case OidDate:
-		return decodeDate(data)
-	case OidTimestamp, OidTimestamptz:
-		return decodeTimestamp(data)
-	case OidTime:
-		return decodeTime(data)
-	case OidInterval:
-		return decodeInterval(data)
-	case OidUUID:
-		return decodeUUID(data)
-	case OidBytea:
-		return data, nil
-	default:
-		// For text, varchar, and unknown types, return as string
-		return string(data), nil
-	}
-}
-
-// formatCopyValue formats a value for COPY output
-func (c *clientConn) formatCopyValue(v interface{}) string {
-	if v == nil {
-		return "\\N"
-	}
-	switch val := v.(type) {
-	case []any:
-		return formatArrayValue(val)
-	case map[string]any:
-		return formatMapValue(val)
-	case arrowmap.OrderedMapValue:
-		return formatOrderedMapValue(val)
-	default:
-		return fmt.Sprintf("%v", val)
-	}
-}
-
-// parseCopyLine parses a line of COPY input
-func (c *clientConn) parseCopyLine(line, delimiter string) []string {
-	// Use encoding/csv for proper handling of quoted values
-	reader := csv.NewReader(strings.NewReader(line))
-	reader.Comma = rune(delimiter[0])
-	reader.LazyQuotes = true // Be lenient with quotes
-
-	fields, err := reader.Read()
-	if err != nil {
-		// Fall back to simple split if CSV parsing fails
-		return strings.Split(line, delimiter)
-	}
-	return fields
-}
-
-func (c *clientConn) sendRowDescription(cols []string, colTypes []ColumnTyper) error {
-	return c.sendRowDescriptionWithFormats(cols, colTypes, nil)
-}
-
-// sendRowDescriptionWithFormats sends a RowDescription message with per-column format codes.
-// formatCodes follow the same convention as Bind result format codes:
-//   - nil or empty: all text (format=0)
-//   - single element: applies to all columns
-//   - one per column: per-column format
-func (c *clientConn) sendRowDescriptionWithFormats(cols []string, colTypes []ColumnTyper, formatCodes []int16) error {
-	var buf bytes.Buffer
-
-	// Number of fields
-	_ = binary.Write(&buf, binary.BigEndian, int16(len(cols)))
-
-	for i, col := range cols {
-		// Strip internal "memory.main." prefix from column names.
-		// In DuckLake mode the transpiler qualifies our custom macros with
-		// memory.main. so DuckDB can find them, but clients shouldn't see that.
-		displayCol := strings.TrimPrefix(col, "memory.main.")
-
-		// Column name (null-terminated)
-		buf.WriteString(displayCol)
-		buf.WriteByte(0)
-
-		// Table OID (0 = not from a table)
-		_ = binary.Write(&buf, binary.BigEndian, int32(0))
-
-		// Column attribute number (0 = not from a table)
-		_ = binary.Write(&buf, binary.BigEndian, int16(0))
-
-		// Data type OID - check for pg_catalog column name overrides first,
-		// then fall back to DuckDB type mapping
-		oid := c.mapTypeOIDWithColumnName(displayCol, colTypes[i])
-		_ = binary.Write(&buf, binary.BigEndian, oid)
-
-		// Data type size - use appropriate size for overridden types
-		typeSize := c.mapTypeSizeWithColumnName(displayCol, colTypes[i])
-		_ = binary.Write(&buf, binary.BigEndian, typeSize)
-
-		// Type modifier (e.g. precision/scale for NUMERIC, -1 = no modifier)
-		typmod := getTypeInfo(colTypes[i]).Typmod
-		_ = binary.Write(&buf, binary.BigEndian, typmod)
-
-		// Format code (0 = text, 1 = binary)
-		var format int16
-		if len(formatCodes) == 1 {
-			format = formatCodes[0]
-		} else if i < len(formatCodes) {
-			format = formatCodes[i]
-		}
-		_ = binary.Write(&buf, binary.BigEndian, format)
-	}
-
-	return wire.WriteMessage(c.writer, wire.MsgRowDescription, buf.Bytes())
-}
-
-func (c *clientConn) mapTypeOIDWithColumnName(colName string, colType ColumnTyper) int32 {
-	// Check if this column name has a specific pg_catalog type override
-	if oid, ok := pgCatalogColumnOIDs[colName]; ok {
-		return oid
-	}
-	return getTypeInfo(colType).OID
-}
-
-func (c *clientConn) mapTypeSizeWithColumnName(colName string, colType ColumnTyper) int16 {
-	// Return appropriate sizes for overridden types
-	if oid, ok := pgCatalogColumnOIDs[colName]; ok {
-		switch oid {
-		case OidName:
-			return 64 // name is 64 bytes
-		case OidChar:
-			return 1 // "char" is 1 byte
-		case OidText:
-			return -1 // text is variable length
-		case OidInt2:
-			return 2 // smallint is 2 bytes
-		}
-	}
-	return getTypeInfo(colType).Size
-}
-
-// sendDataRowWithFormats sends a data row with optional binary encoding
-// formatCodes: per-column format codes (0=text, 1=binary), or nil for all text
-// typeOIDs: per-column type OIDs for binary encoding, or nil
-func (c *clientConn) sendDataRowWithFormats(values []interface{}, formatCodes []int16, typeOIDs []int32) error {
-	var buf bytes.Buffer
-
-	// Number of columns
-	_ = binary.Write(&buf, binary.BigEndian, int16(len(values)))
-
-	for i, v := range values {
-		if v == nil {
-			// NULL value
-			_ = binary.Write(&buf, binary.BigEndian, int32(-1))
-			continue
-		}
-
-		// Determine format: binary or text
-		useBinary := false
-		if formatCodes != nil {
-			if len(formatCodes) == 1 {
-				// Single format code applies to all columns
-				useBinary = formatCodes[0] == 1
-			} else if i < len(formatCodes) {
-				useBinary = formatCodes[i] == 1
-			}
-		}
-
-		if useBinary && typeOIDs != nil && i < len(typeOIDs) {
-			// Binary encoding
-			encoded := encodeBinary(v, typeOIDs[i])
-			if encoded == nil {
-				// Fallback to text if binary encoding fails
-				str := formatValue(v)
-				_ = binary.Write(&buf, binary.BigEndian, int32(len(str)))
-				buf.WriteString(str)
-			} else {
-				_ = binary.Write(&buf, binary.BigEndian, int32(len(encoded)))
-				buf.Write(encoded)
-			}
-		} else {
-			// Text encoding — use JSON re-serialization for JSON columns
-			var str string
-			if typeOIDs != nil && i < len(typeOIDs) && (typeOIDs[i] == OidJSON || typeOIDs[i] == OidJSONB) {
-				str = string(encodeJSON(v))
-			} else {
-				str = formatValue(v)
-			}
-			_ = binary.Write(&buf, binary.BigEndian, int32(len(str)))
-			buf.WriteString(str)
-		}
-	}
-
-	return wire.WriteMessage(c.writer, wire.MsgDataRow, buf.Bytes())
-}
-
-// formatValue converts a value to its PostgreSQL text representation
-func formatValue(v interface{}) string {
-	if v == nil {
-		return ""
-	}
-	v = normalizeDriverValue(v)
-
-	switch val := v.(type) {
-	case []byte:
-		return string(val)
-	case string:
-		return val
-	case *string:
-		if val == nil {
-			return ""
-		}
-		return *val
-	case int, int8, int16, int32, int64:
-		return fmt.Sprintf("%d", val)
-	case uint, uint8, uint16, uint32, uint64:
-		return fmt.Sprintf("%d", val)
-	case float32:
-		return fmt.Sprintf("%g", val)
-	case float64:
-		return fmt.Sprintf("%g", val)
-	case bool:
-		if val {
-			return "t"
-		}
-		return "f"
-	case time.Time:
-		// PostgreSQL timestamp format without timezone suffix
-		if val.IsZero() {
-			return ""
-		}
-		// Use microsecond precision if there are sub-second components
-		if val.Nanosecond() != 0 {
-			return val.Format("2006-01-02 15:04:05.999999")
-		}
-		return val.Format("2006-01-02 15:04:05")
-	case []any:
-		// PostgreSQL array text format: {1,2,3}
-		return formatArrayValue(val)
-	case arrowmap.IntervalValue:
-		// PostgreSQL interval text format: "1 year 2 mons 3 days 04:05:06.123456".
-		// Driver-specific interval types (e.g. duckdb.Interval) are converted to
-		// arrowmap.IntervalValue by normalizeDriverValue before reaching this
-		// switch — see server/value_normalize.go.
-		return formatInterval(val)
-	case map[string]any:
-		// STRUCT text format: {"key1": val1, "key2": val2}
-		return formatMapValue(val)
-	case arrowmap.OrderedMapValue:
-		return formatOrderedMapValue(val)
-	default:
-		// For other types, try to convert to string
-		return fmt.Sprintf("%v", val)
-	}
-}
-
-// formatInterval formats an arrowmap.IntervalValue as a PostgreSQL-compatible
-// interval string. Examples: "00:13:08.917797", "1 day 02:30:00",
-// "1 year 2 mons 3 days 04:05:06".
-func formatInterval(iv arrowmap.IntervalValue) string {
-	var parts []string
-	if iv.Months != 0 {
-		years := iv.Months / 12
-		remMonths := iv.Months % 12
-		if years != 0 {
-			if years == 1 || years == -1 {
-				parts = append(parts, fmt.Sprintf("%d year", years))
-			} else {
-				parts = append(parts, fmt.Sprintf("%d years", years))
-			}
-		}
-		if remMonths != 0 {
-			if remMonths == 1 || remMonths == -1 {
-				parts = append(parts, fmt.Sprintf("%d mon", remMonths))
-			} else {
-				parts = append(parts, fmt.Sprintf("%d mons", remMonths))
-			}
-		}
-	}
-	if iv.Days != 0 {
-		if iv.Days == 1 || iv.Days == -1 {
-			parts = append(parts, fmt.Sprintf("%d day", iv.Days))
-		} else {
-			parts = append(parts, fmt.Sprintf("%d days", iv.Days))
-		}
-	}
-	if iv.Micros != 0 || len(parts) == 0 {
-		micros := iv.Micros
-		neg := micros < 0
-		if neg {
-			micros = -micros
-		}
-		h := micros / 3600000000
-		micros %= 3600000000
-		m := micros / 60000000
-		micros %= 60000000
-		s := micros / 1000000
-		rem := micros % 1000000
-		var timePart string
-		if rem > 0 {
-			timePart = fmt.Sprintf("%02d:%02d:%02d.%06d", h, m, s, rem)
-		} else {
-			timePart = fmt.Sprintf("%02d:%02d:%02d", h, m, s)
-		}
-		if neg {
-			timePart = "-" + timePart
-		}
-		parts = append(parts, timePart)
-	}
-	return strings.Join(parts, " ")
-}
-
-// formatArrayValue formats a []any slice as PostgreSQL text array: {1,2,3}
-func formatArrayValue(arr []any) string {
-	var buf strings.Builder
-	buf.WriteByte('{')
-	for i, elem := range arr {
-		if i > 0 {
-			buf.WriteByte(',')
-		}
-		if elem == nil {
-			buf.WriteString("NULL")
-		} else {
-			s := formatValue(elem)
-			// Quote strings that contain special characters
-			if needsArrayQuoting(s) {
-				buf.WriteByte('"')
-				// Escape backslashes and double quotes
-				for _, c := range s {
-					if c == '"' || c == '\\' {
-						buf.WriteByte('\\')
-					}
-					buf.WriteRune(c)
-				}
-				buf.WriteByte('"')
-			} else {
-				buf.WriteString(s)
-			}
-		}
-	}
-	buf.WriteByte('}')
-	return buf.String()
-}
-
-// needsArrayQuoting returns true if a string value needs quoting inside a PostgreSQL array literal
-func needsArrayQuoting(s string) bool {
-	if s == "" {
-		return true
-	}
-	for _, c := range s {
-		if c == ',' || c == '{' || c == '}' || c == '"' || c == '\\' || c == ' ' {
-			return true
-		}
-	}
-	return false
-}
-
-// formatMapValue formats a map[string]any as a key-value text representation.
-// Used for STRUCT values extracted from Arrow (keys are field names → always strings).
-func formatMapValue(m map[string]any) string {
-	var buf strings.Builder
-	buf.WriteByte('{')
-	first := true
-	for k, v := range m {
-		if !first {
-			buf.WriteString(", ")
-		}
-		first = false
-		buf.WriteString(k)
-		buf.WriteString("=")
-		buf.WriteString(formatValue(v))
-	}
-	buf.WriteByte('}')
-	return buf.String()
-}
-
-// formatOrderedMapValue formats an OrderedMapValue as a key-value text
-// representation, preserving the original insertion order from the Arrow array.
-func formatOrderedMapValue(m arrowmap.OrderedMapValue) string {
-	var buf strings.Builder
-	buf.WriteByte('{')
-	for i, k := range m.Keys {
-		if i > 0 {
-			buf.WriteString(", ")
-		}
-		buf.WriteString(formatValue(k))
-		buf.WriteString("=")
-		buf.WriteString(formatValue(m.Values[i]))
-	}
-	buf.WriteByte('}')
-	return buf.String()
-}
-
-func (c *clientConn) sendError(severity, code, message string) {
-	// Class 28 = "Invalid Authorization Specification" (auth failures).
-	// All current FATAL errors use class 28, so this covers both auth
-	// failures and connection rejections (no SSL, no user, wrong password).
-	// NOTE: If one adds a FATAL error with a non-28 code, be sure to add
-	// a metric for it here.
-	if strings.HasPrefix(code, "28") {
-		auth.AuthFailuresCounter.Inc()
-	} else if severity == "ERROR" {
-		queryErrorsCounter.WithLabelValues(c.orgID).Inc()
-	}
-	slog.Debug("Sending error to client.", "user", c.username, "severity", severity, "code", code, "message", message)
-	c.errorResponsesSent++
-	_ = wire.WriteErrorResponse(c.writer, severity, code, message)
-	_ = c.writer.Flush()
-}
-
-func (c *clientConn) sendNotice(severity, code, message string) {
-	_ = wire.WriteNoticeResponse(c.writer, severity, code, message)
-	// Don't flush here - let the caller decide when to flush
-}
-
 // Extended query protocol handlers
-
-func (c *clientConn) handleParse(body []byte) {
-	// Parse message format:
-	// - Statement name (null-terminated string)
-	// - Query string (null-terminated string)
-	// - Number of parameter types (int16)
-	// - Parameter type OIDs (int32 each)
-
-	reader := bytes.NewReader(body)
-
-	// Read statement name
-	stmtName, err := readCString(reader)
-	if err != nil {
-		c.sendError("ERROR", "08P01", "invalid Parse message")
-		return
-	}
-
-	// Read query
-	query, err := readCString(reader)
-	if err != nil {
-		c.sendError("ERROR", "08P01", "invalid Parse message")
-		return
-	}
-	// Read number of parameter types
-	var numParamTypes int16
-	if err := binary.Read(reader, binary.BigEndian, &numParamTypes); err != nil {
-		c.sendError("ERROR", "08P01", "invalid Parse message")
-		return
-	}
-
-	// Read parameter type OIDs
-	paramTypes := make([]int32, numParamTypes)
-	for i := int16(0); i < numParamTypes; i++ {
-		if err := binary.Read(reader, binary.BigEndian, &paramTypes[i]); err != nil {
-			c.sendError("ERROR", "08P01", "invalid Parse message")
-			return
-		}
-	}
-
-	// Detect cursor operations before passthrough or transpilation.
-	// DuckDB doesn't support DECLARE/FETCH/CLOSE natively, so cursor
-	// emulation is needed for all users including passthrough.
-	cursorTree, cursorParseErr := pg_query.Parse(query)
-	if cursorParseErr == nil && len(cursorTree.Stmts) == 1 {
-		switch s := cursorTree.Stmts[0].Stmt.Node.(type) {
-		case *pg_query.Node_DeclareCursorStmt:
-			innerSQL := deparseInnerQuery(s.DeclareCursorStmt.Query)
-			transpiledSQL := innerSQL
-			if !c.passthrough && innerSQL != "" {
-				tr := c.newTranspiler(true)
-				innerResult, innerErr := tr.Transpile(innerSQL)
-				if innerErr == nil && !innerResult.FallbackToNative {
-					transpiledSQL = innerResult.SQL
-				}
-			}
-			delete(c.stmts, stmtName)
-			c.stmts[stmtName] = &preparedStmt{
-				query:          query,
-				convertedQuery: query,
-				cursorOp:       cursorOpDeclare,
-				cursorName:     s.DeclareCursorStmt.Portalname,
-				cursorQuery:    transpiledSQL,
-			}
-			_ = wire.WriteParseComplete(c.writer)
-			return
-
-		case *pg_query.Node_FetchStmt:
-			if !isFetchForwardOnly(s.FetchStmt.Direction) || s.FetchStmt.HowMany < 0 {
-				c.sendError("ERROR", "0A000", "cursor can only scan forward")
-				return
-			}
-			delete(c.stmts, stmtName)
-			c.stmts[stmtName] = &preparedStmt{
-				query:          query,
-				convertedQuery: query,
-				cursorOp:       cursorOpFetch,
-				cursorName:     s.FetchStmt.Portalname,
-				fetchCount:     s.FetchStmt.HowMany,
-			}
-			_ = wire.WriteParseComplete(c.writer)
-			return
-
-		case *pg_query.Node_ClosePortalStmt:
-			delete(c.stmts, stmtName)
-			c.stmts[stmtName] = &preparedStmt{
-				query:          query,
-				convertedQuery: query,
-				cursorOp:       cursorOpClose,
-				cursorName:     s.ClosePortalStmt.Portalname,
-			}
-			_ = wire.WriteParseComplete(c.writer)
-			return
-		}
-	}
-
-	// Intercept pg_cursors queries (e.g. psycopg's "SELECT 1 FROM pg_cursors WHERE name = $1").
-	// DuckDB doesn't have this system view; return synthetic results from cursor emulation state.
-	if cursorName, parameterized, ok := matchPgCursorsQuery(query); ok {
-		delete(c.stmts, stmtName)
-		ps := &preparedStmt{
-			query:          query,
-			convertedQuery: query,
-			cursorOp:       cursorOpPgCursorsQuery,
-			cursorName:     cursorName,
-		}
-		if parameterized {
-			ps.numParams = 1
-			ps.paramTypes = []int32{25} // text OID
-		}
-		c.stmts[stmtName] = ps
-		_ = wire.WriteParseComplete(c.writer)
-		return
-	}
-
-	// Intercept pg_stat_activity queries. Return synthetic results from the connection registry.
-	if matchPgStatActivityQuery(query) {
-		delete(c.stmts, stmtName)
-		c.stmts[stmtName] = &preparedStmt{
-			query:          query,
-			convertedQuery: query,
-			cursorOp:       cursorOpPgStatActivity,
-		}
-		_ = wire.WriteParseComplete(c.writer)
-		return
-	}
-
-	// Passthrough mode: skip transpilation, store query directly
-	if c.passthrough {
-		// Count $N parameters with a simple regex (pg_query.Parse may fail on DuckDB-native SQL)
-		paramCount := countDollarParams(query)
-		delete(c.stmts, stmtName)
-		c.stmts[stmtName] = &preparedStmt{
-			query:          query,
-			convertedQuery: query, // No transpilation
-			paramTypes:     paramTypes,
-			numParams:      paramCount,
-		}
-		_ = wire.WriteParseComplete(c.writer)
-		return
-	}
-
-	// Transpile PostgreSQL SQL to DuckDB-compatible SQL (with placeholder conversion)
-	tr := c.newTranspiler(true) // Enable placeholder conversion for prepared statements
-	result, err := tr.Transpile(query)
-	if err != nil {
-		c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-		return
-	}
-
-	// Handle transform-detected errors (e.g., unrecognized config parameter)
-	if result.Error != nil {
-		c.sendError("ERROR", transformErrorSQLState(result.Error), result.Error.Error())
-		return
-	}
-
-	// Handle fallback to native DuckDB: PostgreSQL parsing failed, try DuckDB directly
-	if result.FallbackToNative {
-		if err := c.validateWithDuckDB(query); err != nil {
-			// Neither PostgreSQL nor DuckDB can parse this query
-			c.sendError("ERROR", "42601", fmt.Sprintf("syntax error: %v", err))
-			return
-		}
-		slog.Debug("Fallback to native DuckDB: query not valid PostgreSQL but valid DuckDB.", "user", c.username, "query", query)
-	}
-
-	// Close existing statement with same name
-	delete(c.stmts, stmtName)
-
-	c.stmts[stmtName] = &preparedStmt{
-		query:             query,                            // Keep original for logging and Describe
-		convertedQuery:    c.rewriteDirectQuery(result.SQL), // Transpiled SQL for execution
-		paramTypes:        paramTypes,
-		numParams:         result.ParamCount,
-		isIgnoredSet:      result.IsIgnoredSet,
-		isNoOp:            result.IsNoOp,
-		noOpTag:           result.NoOpTag,
-		statements:        result.Statements,        // Multi-statement rewrite (writable CTE)
-		cleanupStatements: result.CleanupStatements, // Cleanup statements
-		warnings:          result.Warnings,          // Surfaced as NoticeResponse at Execute
-	}
-
-	slog.Debug("Prepared statement.", "user", c.username, "name", stmtName, "query", query)
-	if len(result.Statements) > 0 {
-		slog.Debug("Prepared statement multi-statement.", "user", c.username, "name", stmtName, "statements", len(result.Statements), "cleanup", len(result.CleanupStatements))
-	} else if result.SQL != query {
-		slog.Debug("Prepared statement transpiled.", "user", c.username, "name", stmtName, "transpiled", result.SQL)
-	}
-	_ = wire.WriteParseComplete(c.writer)
-}
 
 func (c *clientConn) handleBind(body []byte) {
 	// Bind message format:
@@ -5575,1373 +2186,4 @@ func (c *clientConn) handleBind(body []byte) {
 	}
 
 	_ = wire.WriteBindComplete(c.writer)
-}
-
-func (c *clientConn) handleDescribe(body []byte) {
-	// Describe message format:
-	// - Type: 'S' for statement, 'P' for portal
-	// - Name (null-terminated)
-
-	if len(body) < 2 {
-		c.sendError("ERROR", "08P01", "invalid Describe message")
-		return
-	}
-
-	descType := body[0]
-	name := string(bytes.TrimRight(body[1:], "\x00"))
-
-	switch descType {
-	case 'S':
-		// Describe prepared statement
-		ps, ok := c.stmts[name]
-		if !ok {
-			c.sendError("ERROR", "26000", fmt.Sprintf("prepared statement %q does not exist", name))
-			return
-		}
-		slog.Debug("Describe statement.", "user", c.username, "name", name, "query", ps.query)
-
-		// Send parameter description based on the number of $N placeholders we found
-		// If the client didn't send explicit types, create them
-		paramTypes := ps.paramTypes
-		if len(paramTypes) < ps.numParams {
-			paramTypes = make([]int32, ps.numParams)
-			// Default to text type for unspecified params
-			for i := range paramTypes {
-				paramTypes[i] = 25 // text OID
-			}
-		}
-		c.sendParameterDescription(paramTypes)
-
-		// Handle cursor operations in Describe
-		switch ps.cursorOp {
-		case cursorOpDeclare, cursorOpClose:
-			// DECLARE and CLOSE don't return rows
-			_ = wire.WriteNoData(c.writer)
-			return
-		case cursorOpFetch:
-			// FETCH returns rows — look up cursor to get schema
-			cols, colTypes, err := c.getCursorSchema(ps.cursorName)
-			if err != nil || len(cols) == 0 {
-				_ = wire.WriteNoData(c.writer)
-				return
-			}
-			_ = c.sendRowDescription(cols, colTypes)
-			ps.described = true
-			return
-		case cursorOpPgCursorsQuery:
-			_ = c.sendPgCursorsRowDescriptionWithFormats(nil)
-			ps.described = true
-			return
-		case cursorOpPgStatActivity:
-			_ = c.sendPgStatActivityRowDescriptionWithFormats(nil)
-			ps.described = true
-			return
-		}
-
-		// For queries that return results, we need to send RowDescription
-		// For other queries, send NoData
-		returnsResults := queryReturnsResults(ps.query)
-		slog.Debug("Describe statement returns results check.", "user", c.username, "name", name, "returns_results", returnsResults)
-		if !returnsResults {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		// DML with RETURNING cannot be described without executing the mutation.
-		// Reject with an explicit error so clients don't desync (e.g., lib/pq
-		// would use Exec-like handling after NoData, silently dropping rows).
-		if isDMLReturning(ps.query) {
-			c.sendError("ERROR", "0A000", "DML with RETURNING clause cannot be described without executing the mutation; use simple query protocol or skip the Describe step")
-			return
-		}
-
-		// WITH + DML (no RETURNING) doesn't return results but queryReturnsResults
-		// returns true for all WITH-prefixed queries. Send NoData to avoid executing
-		// the mutation during schema probing.
-		if isWithDML(ps.query) {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		// EXPLAIN [ANALYZE] returns a single textual plan column. Describing it via
-		// the LIMIT-0 probe below would EXECUTE it — and EXPLAIN ANALYZE of a write
-		// mutates — so the statement would run at Describe and again at Execute.
-		// Send a synthetic RowDescription without executing.
-		if isExplainStmt(ps.query) {
-			_ = c.sendRowDescription([]string{explainPlanColumn(ps.query)}, []ColumnTyper{staticColumnType("VARCHAR")})
-			ps.described = true
-			return
-		}
-
-		// For SELECT, we need to describe the result columns
-		// The cleanest approach is to add a "WHERE false" or "LIMIT 0" clause
-		// to get column info without actually running the query
-		describeQuery := strings.TrimRight(strings.TrimSpace(ps.convertedQuery), ";")
-		// Try adding LIMIT 0 to avoid needing real parameter values.
-		// Only for statements that support LIMIT (SELECT/WITH/VALUES/TABLE/FROM).
-		upperDesc := strings.ToUpper(describeQuery)
-		if !strings.Contains(upperDesc, "LIMIT") && describeSupportsLimit(upperDesc) {
-			describeQuery = describeQuery + " LIMIT 0"
-		}
-
-		// Use NULL for all parameters
-		args := make([]interface{}, ps.numParams)
-		for i := range args {
-			args[i] = nil
-		}
-
-		rows, err := c.executor.Query(describeQuery, args...)
-		if err != nil {
-			// Can't describe - send NoData
-			slog.Debug("Describe failed to get columns.", "user", c.username, "error", err)
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		cols, _ := rows.Columns()
-		colTypes, _ := rows.ColumnTypes()
-		_ = rows.Close()
-
-		if len(cols) == 0 {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		slog.Debug("Describe statement sending RowDescription.", "user", c.username, "columns", len(cols))
-		_ = c.sendRowDescription(cols, colTypes)
-		ps.described = true
-
-	case 'P':
-		// Describe portal
-		p, ok := c.portals[name]
-		if !ok {
-			// In PostgreSQL, DECLARE CURSOR creates a named cursor that is also
-			// accessible as a portal. psycopg3's ServerCursor sends Describe Portal
-			// with the cursor name after DECLARE. Check c.cursors as fallback.
-			if _, cursorOk := c.cursors[name]; cursorOk {
-				cols, colTypes, err := c.getCursorSchema(name)
-				if err != nil {
-					slog.Debug("Describe cursor-as-portal failed to open.", "user", c.username, "cursor", name, "error", err)
-					_ = wire.WriteNoData(c.writer)
-					return
-				}
-				_ = c.sendRowDescription(cols, colTypes)
-				return
-			}
-			c.sendError("ERROR", "34000", fmt.Sprintf("portal %q does not exist", name))
-			return
-		}
-
-		// Handle cursor operations in portal Describe
-		switch p.stmt.cursorOp {
-		case cursorOpDeclare, cursorOpClose:
-			_ = wire.WriteNoData(c.writer)
-			return
-		case cursorOpFetch:
-			cols, colTypes, err := c.getCursorSchema(p.stmt.cursorName)
-			if err != nil || len(cols) == 0 {
-				_ = wire.WriteNoData(c.writer)
-				return
-			}
-			p.described = true
-			_ = c.sendRowDescriptionWithFormats(cols, colTypes, p.resultFormats)
-			return
-		case cursorOpPgCursorsQuery:
-			_ = c.sendPgCursorsRowDescriptionWithFormats(p.resultFormats)
-			p.described = true
-			return
-		case cursorOpPgStatActivity:
-			_ = c.sendPgStatActivityRowDescriptionWithFormats(p.resultFormats)
-			p.described = true
-			return
-		}
-
-		// For queries that don't return results, send NoData
-		if !queryReturnsResults(p.stmt.query) {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		// DML with RETURNING cannot be described without executing the mutation.
-		// Reject with an explicit error so clients don't desync.
-		if isDMLReturning(p.stmt.query) {
-			c.sendError("ERROR", "0A000", "DML with RETURNING clause cannot be described without executing the mutation; use simple query protocol or skip the Describe step")
-			return
-		}
-
-		// WITH + DML (no RETURNING) doesn't return results but queryReturnsResults
-		// returns true for all WITH-prefixed queries. Send NoData to avoid executing
-		// the mutation during schema probing.
-		if isWithDML(p.stmt.query) {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		// EXPLAIN [ANALYZE]: synthesize the single plan column without executing
-		// (see the statement-Describe branch above).
-		if isExplainStmt(p.stmt.query) {
-			_ = c.sendRowDescriptionWithFormats([]string{explainPlanColumn(p.stmt.query)}, []ColumnTyper{staticColumnType("VARCHAR")}, p.resultFormats)
-			p.described = true
-			p.stmt.described = true
-			return
-		}
-
-		// For SELECT, we need to describe the result columns
-		// We'll do a trial query with LIMIT 0 to get column info
-		args, err := p.decodeParams()
-		if err != nil {
-			// PostgreSQL returns 08P01 (protocol violation) for malformed binary data
-			c.sendError("ERROR", "08P01", fmt.Sprintf("insufficient data left in message: %v", err))
-			return
-		}
-
-		// Try to get column info without fully executing expensive queries.
-		describeQuery := strings.TrimRight(strings.TrimSpace(p.stmt.convertedQuery), ";")
-		upperDesc := strings.ToUpper(describeQuery)
-		if !strings.Contains(upperDesc, "LIMIT") && describeSupportsLimit(upperDesc) {
-			describeQuery = describeQuery + " LIMIT 0"
-		}
-
-		rows, err := c.executor.Query(describeQuery, args...)
-		if err != nil {
-			// Can't describe - send NoData
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		cols, _ := rows.Columns()
-		colTypes, _ := rows.ColumnTypes()
-		_ = rows.Close()
-
-		if len(cols) == 0 {
-			_ = wire.WriteNoData(c.writer)
-			return
-		}
-
-		// Mark both portal and statement as described when we send RowDescription.
-		// If we sent NoData above, Execute should still send RowDescription.
-		// Setting ps.described ensures future Bind calls that create new portals
-		// from this statement inherit described=true, so Execute won't re-send
-		// RowDescription. Without this, JDBC drivers that reuse named statements
-		// (Bind/Execute without re-Describing) get an unexpected RowDescription
-		// and desync their message queue.
-		p.described = true
-		p.stmt.described = true
-		_ = c.sendRowDescriptionWithFormats(cols, colTypes, p.resultFormats)
-
-	default:
-		c.sendError("ERROR", "08P01", "invalid Describe type")
-	}
-}
-
-func (c *clientConn) handleExecute(body []byte) {
-	// Execute message format:
-	// - Portal name (null-terminated)
-	// - Maximum rows to return (int32, 0 = no limit)
-
-	reader := bytes.NewReader(body)
-
-	portalName, err := readCString(reader)
-	if err != nil {
-		c.sendError("ERROR", "08P01", "invalid Execute message")
-		return
-	}
-
-	var maxRows int32
-	if err := binary.Read(reader, binary.BigEndian, &maxRows); err != nil {
-		c.sendError("ERROR", "08P01", "invalid Execute message")
-		return
-	}
-
-	p, ok := c.portals[portalName]
-	if !ok {
-		c.sendError("ERROR", "34000", fmt.Sprintf("portal %q does not exist", portalName))
-		return
-	}
-
-	c.currentQuery.Store(p.stmt.query)
-	c.queryStart.Store(time.Now())
-	defer func() {
-		c.currentQuery.Store("")
-		c.queryStart.Store(time.Time{})
-	}()
-
-	// Handle cursor operations before normal execution
-	switch p.stmt.cursorOp {
-	case cursorOpDeclare:
-		c.handleDeclareCursorExtended(p)
-		return
-	case cursorOpFetch:
-		c.handleFetchCursorExtended(p)
-		return
-	case cursorOpClose:
-		c.handleCloseCursorExtended(p)
-		return
-	case cursorOpPgCursorsQuery:
-		c.handlePgCursorsQueryExtended(p)
-		return
-	case cursorOpPgStatActivity:
-		c.handlePgStatActivityExtended(p)
-		return
-	}
-
-	// Handle empty queries - PostgreSQL returns EmptyQueryResponse for these
-	trimmedQuery := strings.TrimSpace(p.stmt.query)
-	if trimmedQuery == "" || isEmptyQuery(trimmedQuery) {
-		_ = wire.WriteEmptyQueryResponse(c.writer)
-		return
-	}
-
-	start := time.Now()
-	defer func() { queryDurationHistogram.WithLabelValues(c.orgID).Observe(time.Since(start).Seconds()) }()
-
-	queryCtx, span := observe.Tracer().Start(c.ctx, "duckgres.query",
-		trace.WithAttributes(
-			attribute.String("duckgres.protocol", "extended"),
-			attribute.String("duckgres.org_id", c.orgID),
-			attribute.String("db.user", c.username),
-			attribute.String("db.statement", observe.TruncateForSpan(p.stmt.query)),
-		),
-	)
-	defer span.End()
-
-	// Convert parameter values to interface{}, handling binary format
-	args, err := p.decodeParams()
-	if err != nil {
-		// PostgreSQL returns 08P01 (protocol violation) for malformed binary data
-		c.sendError("ERROR", "08P01", fmt.Sprintf("insufficient data left in message: %v", err))
-		return
-	}
-
-	upperQuery := strings.ToUpper(strings.TrimSpace(p.stmt.query))
-	cmdType := c.getCommandType(upperQuery)
-	returnsResults := queryReturnsResults(p.stmt.query)
-
-	slog.Debug("Execute portal.", "user", c.username, "portal", portalName, "params", len(args), "query", p.stmt.query)
-
-	// Surface any transpiler warnings (e.g. an unenforced constraint stripped on a
-	// lake catalog) as NoticeResponse before the command result.
-	for _, w := range p.stmt.warnings {
-		c.sendNotice("WARNING", "01000", w)
-	}
-
-	// Check if this is a PostgreSQL-specific SET command that should be ignored
-	// (determined by transpiler during Parse)
-	if p.stmt.isIgnoredSet {
-		slog.Debug("Ignoring PostgreSQL-specific SET.", "user", c.username, "query", p.stmt.query)
-		_ = wire.WriteCommandComplete(c.writer, "SET")
-		return
-	}
-
-	// Handle no-op commands (CREATE INDEX, VACUUM, etc.) - DuckLake doesn't support these
-	// (determined by transpiler during Parse)
-	if p.stmt.isNoOp {
-		slog.Debug("No-op command (DuckLake limitation).", "user", c.username, "query", p.stmt.query)
-		_ = wire.WriteCommandComplete(c.writer, p.stmt.noOpTag)
-		return
-	}
-
-	// Handle multi-statement results (e.g., writable CTE rewrites)
-	if len(p.stmt.statements) > 0 {
-		slog.Debug("Execute multi-statement.", "user", c.username, "statements", len(p.stmt.statements), "cleanup", len(p.stmt.cleanupStatements))
-		c.executeMultiStatementExtended(p.stmt.statements, p.stmt.cleanupStatements, args, p.resultFormats, p.described)
-		return
-	}
-
-	originalQuery := p.stmt.query
-	convertedQuery := p.stmt.convertedQuery
-
-	// Lifecycle log pair for the extended-query path. logQueryStarted /
-	// logQueryFinished are the canonical "did a query run on a worker?"
-	// signal for Loki / Grafana (PR #519). Without these, the only log a
-	// successful extended-query produces is the structured logQuery() to
-	// the queryLogger channel, which doesn't carry worker_id and isn't
-	// scrape-friendly. queryFinalErr is captured by the deferred call so
-	// every termination path — success, ALTER-TABLE-as-VIEW retry,
-	// transaction-conflict retry, recovery rollback, fatal error — emits
-	// exactly one Finished log per Started.
-	queryStart := time.Now()
-	var queryRowsAff int64
-	var queryFinalErr error
-	c.logQueryStarted(convertedQuery)
-	defer func() {
-		c.logQueryFinished(convertedQuery, queryStart, queryRowsAff, queryFinalErr)
-	}()
-
-	if !returnsResults {
-		// Handle nested BEGIN: PostgreSQL issues a warning but continues,
-		// while DuckDB throws an error. Match PostgreSQL behavior.
-		if cmdType == "BEGIN" && c.txStatus == txStatusTransaction {
-			c.sendNotice("WARNING", "25001", "there is already a transaction in progress")
-			_ = wire.WriteCommandComplete(c.writer, "BEGIN")
-			return
-		}
-
-		// Non-result-returning query: use Exec with converted query
-		runExec := func() (ExecResult, error) {
-			result, err := c.executor.Exec(convertedQuery, args...)
-			if err != nil {
-				if fallbackResult, handled, fallbackErr := c.execCompatibilityFallback(queryCtx, convertedQuery, err, func(fallbackQuery string) (ExecResult, error) {
-					return c.executor.Exec(fallbackQuery, args...)
-				}); handled {
-					return fallbackResult, fallbackErr
-				}
-			}
-			return result, err
-		}
-
-		execStart := time.Now()
-		execCtx, execSpan := observe.Tracer().Start(queryCtx, "duckgres.execute")
-		result, err := runExec()
-		c.lastProfilingSummary = observe.EnrichSpanWithProfiling(execCtx, execSpan, execStart, c.executor, c.orgID)
-		execSpan.End()
-		if err != nil {
-			if c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-				ducklakeConflictTotal.Inc()
-				result, err = retryOnConflict(runExec)
-			}
-			if err != nil {
-				result, err, _ = recoverAbortedTransaction(
-					err,
-					c.txStatus == txStatusIdle,
-					func() error {
-						_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-						return rollbackErr
-					},
-					runExec,
-				)
-			}
-			if err != nil {
-				queryFinalErr = err
-				errCode := classifyErrorCode(err)
-				errMsg := friendlyExecError(err)
-				if c.isCallerCancellation(err) {
-					errMsg = "canceling statement due to user request"
-				} else {
-					c.logQueryError(convertedQuery, err)
-				}
-				c.sendError("ERROR", errCode, errMsg)
-				c.setTxError()
-				c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, 0, errCode, errMsg, "extended")
-				return
-			}
-		}
-		var writtenRows int64
-		if result != nil {
-			writtenRows, _ = result.RowsAffected()
-		}
-		queryRowsAff = writtenRows
-		c.updateTxStatus(cmdType)
-		tag := c.buildCommandTag(cmdType, result)
-		_ = wire.WriteCommandComplete(c.writer, tag)
-		c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, writtenRows, "", "", "extended")
-		return
-	}
-
-	// Result-returning query: use Query with converted query
-	runQuery := func() (RowSet, error) {
-		return c.queryWithArgsWithMetadata(queryCtx, convertedQuery, args...)
-	}
-
-	execStart := time.Now()
-	execCtx, execSpan := observe.Tracer().Start(queryCtx, "duckgres.execute")
-	rows, err := runQuery()
-	if err != nil && c.txStatus == txStatusIdle && isDuckLakeTransactionConflict(err) {
-		ducklakeConflictTotal.Inc()
-		rows, err = retryOnConflict(runQuery)
-	}
-	if err != nil {
-		rows, err, _ = recoverAbortedTransaction(
-			err,
-			c.txStatus == txStatusIdle,
-			func() error {
-				_, rollbackErr := c.executor.ExecContext(context.Background(), "ROLLBACK")
-				return rollbackErr
-			},
-			runQuery,
-		)
-	}
-	c.lastProfilingSummary = observe.EnrichSpanWithProfiling(execCtx, execSpan, execStart, c.executor, c.orgID)
-	execSpan.End()
-	if err != nil {
-		queryFinalErr = err
-		errCode := classifyErrorCode(err)
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errMsg = "canceling statement due to user request"
-		} else {
-			c.logQueryError(convertedQuery, err)
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, 0, errCode, errMsg, "extended")
-		return
-	}
-	defer func() { _ = rows.Close() }()
-
-	cols, err := rows.Columns()
-	if err != nil {
-		queryFinalErr = err
-		slog.Error("Columns error.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-		c.sendError("ERROR", "42000", err.Error())
-		c.setTxError()
-		c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, 0, "42000", err.Error(), "extended")
-		return
-	}
-
-	// Get column types for binary encoding
-	colTypes, _ := rows.ColumnTypes()
-	typeOIDs := make([]int32, len(cols))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-
-	// Send RowDescription if Describe wasn't called before Execute.
-	// Some clients skip Describe and go straight to Execute, but still
-	// need the column metadata before receiving data rows.
-	// Skip if there are no columns - queries that return 0 columns (like
-	// DDL accidentally routed here) don't need RowDescription.
-	if !p.described && len(cols) > 0 {
-		if err := c.sendRowDescriptionWithFormats(cols, colTypes, p.resultFormats); err != nil {
-			return
-		}
-	}
-
-	// Send rows with the format codes from Bind
-	rowCount := 0
-	for rows.Next() {
-		if maxRows > 0 && int32(rowCount) >= maxRows {
-			// Portal suspended - but we don't support this yet
-			break
-		}
-
-		values := make([]interface{}, len(cols))
-		valuePtrs := make([]interface{}, len(cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := rows.Scan(valuePtrs...); err != nil {
-			queryFinalErr = err
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, 0, "42000", err.Error(), "extended")
-			return
-		}
-
-		if err := c.sendDataRowWithFormats(values, p.resultFormats, typeOIDs); err != nil {
-			queryFinalErr = err
-			return
-		}
-		rowCount++
-	}
-	queryRowsAff = int64(rowCount)
-
-	if err := rows.Err(); err != nil {
-		queryFinalErr = err
-		errCode := "42000"
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errCode = "57014"
-			errMsg = "canceling statement due to user request"
-			c.sendError("ERROR", errCode, errMsg)
-		} else {
-			slog.Error("Row iteration error.", "user", c.username, "error", err, "worker", c.workerID, "worker_pod", c.workerPod)
-			c.sendError("ERROR", errCode, errMsg)
-		}
-		c.setTxError()
-		c.logQuery(start, originalQuery, convertedQuery, cmdType, 0, 0, errCode, errMsg, "extended")
-		return
-	}
-
-	c.updateTxStatus(cmdType)
-	tag := buildCommandTagFromRowCount(cmdType, int64(rowCount))
-	_ = wire.WriteCommandComplete(c.writer, tag)
-	c.logQuery(start, originalQuery, convertedQuery, cmdType, int64(rowCount), 0, "", "", "extended")
-}
-
-func (c *clientConn) handleClose(body []byte) {
-	// Close message format:
-	// - Type: 'S' for statement, 'P' for portal
-	// - Name (null-terminated)
-
-	if len(body) < 2 {
-		c.sendError("ERROR", "08P01", "invalid Close message")
-		return
-	}
-
-	closeType := body[0]
-	name := string(bytes.TrimRight(body[1:], "\x00"))
-
-	switch closeType {
-	case 'S':
-		delete(c.stmts, name)
-	case 'P':
-		delete(c.portals, name)
-	}
-
-	_ = wire.WriteCloseComplete(c.writer)
-}
-
-func (c *clientConn) sendParameterDescription(paramTypes []int32) {
-	var buf bytes.Buffer
-	_ = binary.Write(&buf, binary.BigEndian, int16(len(paramTypes)))
-	for _, oid := range paramTypes {
-		// If OID is 0, use text type
-		if oid == 0 {
-			oid = 25 // text
-		}
-		_ = binary.Write(&buf, binary.BigEndian, oid)
-	}
-	_ = wire.WriteMessage(c.writer, 't', buf.Bytes())
-}
-
-// readCString reads a null-terminated string from reader
-func readCString(r *bytes.Reader) (string, error) {
-	var buf bytes.Buffer
-	for {
-		b, err := r.ReadByte()
-		if err != nil {
-			return "", err
-		}
-		if b == 0 {
-			break
-		}
-		buf.WriteByte(b)
-	}
-	return buf.String(), nil
-}
-
-// --- Server-side cursor emulation ---
-
-// deparseInnerQuery deparses a pg_query Node back to SQL text.
-func deparseInnerQuery(node *pg_query.Node) string {
-	tree := &pg_query.ParseResult{
-		Stmts: []*pg_query.RawStmt{{Stmt: node}},
-	}
-	sql, err := pg_query.Deparse(tree)
-	if err != nil {
-		return ""
-	}
-	return sql
-}
-
-// openCursor executes the cursor's stored query and caches the result set metadata.
-func (c *clientConn) openCursor(cursor *cursorState) error {
-	ctx, cleanup := c.queryContextForCursor()
-	cursor.cleanup = cleanup
-
-	// Lifecycle log pair (PR #519): the cursor's underlying SELECT is the
-	// query the worker runs at OPEN time. The eventual row count comes
-	// from later FETCH iterations against the same rowset, which the
-	// outer cursor lifecycle (closeCursor) doesn't currently bubble up to
-	// us — log the cursor metadata phase as rows=0 / err=initial failure.
-	cursorStart := time.Now()
-	c.logQueryStarted(cursor.query)
-	rows, err := c.queryContextWithMetadata(ctx, cursor.query)
-	if err != nil {
-		c.logQueryFinished(cursor.query, cursorStart, 0, err)
-		cleanup()
-		cursor.cleanup = nil
-		return err
-	}
-	cursor.rows = rows
-
-	cols, err := rows.Columns()
-	if err != nil {
-		c.logQueryFinished(cursor.query, cursorStart, 0, err)
-		_ = rows.Close()
-		cursor.rows = nil
-		cleanup()
-		cursor.cleanup = nil
-		return err
-	}
-	cursor.cols = cols
-
-	colTypes, err := rows.ColumnTypes()
-	if err != nil {
-		c.logQueryFinished(cursor.query, cursorStart, 0, err)
-		_ = rows.Close()
-		cursor.rows = nil
-		cleanup()
-		cursor.cleanup = nil
-		return err
-	}
-	cursor.colTypes = colTypes
-	// Cursor opened successfully — emit the matching Finished now (rows=0,
-	// since rows are streamed via FETCH later and we'd otherwise log
-	// nothing on the success path).
-	c.logQueryFinished(cursor.query, cursorStart, 0, nil)
-
-	typeOIDs := make([]int32, len(colTypes))
-	for i, ct := range colTypes {
-		typeOIDs[i] = getTypeInfo(ct).OID
-	}
-	cursor.typeOIDs = typeOIDs
-	return nil
-}
-
-// closeCursor closes a specific cursor and cleans up its resources.
-func (c *clientConn) closeCursor(name string) {
-	cursor, ok := c.cursors[name]
-	if !ok {
-		return
-	}
-	if cursor.rows != nil {
-		_ = cursor.rows.Close()
-	}
-	if cursor.cleanup != nil {
-		cursor.cleanup()
-	}
-	delete(c.cursors, name)
-}
-
-// closeAllCursors closes all open cursors on this connection.
-func (c *clientConn) closeAllCursors() {
-	for name := range c.cursors {
-		c.closeCursor(name)
-	}
-}
-
-// getCursorSchema opens the cursor if needed to retrieve column metadata,
-// then returns the schema information. Used by handleDescribe for FETCH statements.
-func (c *clientConn) getCursorSchema(cursorName string) ([]string, []ColumnTyper, error) {
-	cursor, ok := c.cursors[cursorName]
-	if !ok {
-		return nil, nil, fmt.Errorf("cursor %q does not exist", cursorName)
-	}
-	if cursor.rows == nil {
-		if err := c.openCursor(cursor); err != nil {
-			return nil, nil, err
-		}
-	}
-	return cursor.cols, cursor.colTypes, nil
-}
-
-// isFetchForwardOnly returns true if the FetchStmt direction is forward-compatible.
-func isFetchForwardOnly(dir pg_query.FetchDirection) bool {
-	return dir == pg_query.FetchDirection_FETCH_DIRECTION_UNDEFINED ||
-		dir == pg_query.FetchDirection_FETCH_FORWARD
-}
-
-// pgCursorsLiteralRegex matches pg_cursors queries with a literal name value:
-//
-//	SELECT 1 FROM pg_cursors WHERE name = 'cursor_name'
-//	SELECT 1 FROM pg_catalog.pg_cursors WHERE name = 'cursor_name'
-var pgCursorsLiteralRegex = regexp.MustCompile(
-	`(?i)^\s*SELECT\s+.+\s+FROM\s+(?:pg_catalog\s*\.\s*)?pg_cursors\s+WHERE\s+name\s*=\s*'([^']*)'`,
-)
-
-// pgCursorsParamRegex matches pg_cursors queries with a parameterized name:
-//
-//	SELECT 1 FROM pg_cursors WHERE name = $1
-var pgCursorsParamRegex = regexp.MustCompile(
-	`(?i)^\s*SELECT\s+.+\s+FROM\s+(?:pg_catalog\s*\.\s*)?pg_cursors\s+WHERE\s+name\s*=\s*\$1\s*$`,
-)
-
-// matchPgCursorsQuery checks if a query is a pg_cursors lookup and extracts the cursor name.
-// Returns the cursor name and true for literal queries, empty string and true for parameterized queries.
-func matchPgCursorsQuery(query string) (cursorName string, parameterized bool, ok bool) {
-	if !strings.Contains(query, "pg_cursors") {
-		return "", false, false
-	}
-	if m := pgCursorsLiteralRegex.FindStringSubmatch(query); m != nil {
-		return m[1], false, true
-	}
-	if pgCursorsParamRegex.MatchString(query) {
-		return "", true, true
-	}
-	return "", false, false
-}
-
-// handlePgCursorsQuery handles SELECT FROM pg_cursors in the Simple Query protocol.
-// Returns a single row with value "1" if the cursor exists, or zero rows if not.
-func (c *clientConn) handlePgCursorsQuery(cursorName string) error {
-	_, exists := c.cursors[cursorName]
-
-	// Send RowDescription: single int4 column named "?column?"
-	if err := c.sendPgCursorsRowDescriptionWithFormats(nil); err != nil {
-		return err
-	}
-
-	rowCount := 0
-	if exists {
-		if err := c.sendDataRowWithFormats([]interface{}{int64(1)}, nil, []int32{23}); err != nil {
-			return err
-		}
-		rowCount = 1
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", rowCount))
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handlePgCursorsQueryExtended handles SELECT FROM pg_cursors in the Extended Query protocol.
-func (c *clientConn) handlePgCursorsQueryExtended(p *portal) {
-	// Resolve cursor name: either from literal in query or from bind parameter
-	cursorName := p.stmt.cursorName
-	if cursorName == "" && len(p.paramValues) > 0 && p.paramValues[0] != nil {
-		cursorName = string(p.paramValues[0])
-	}
-
-	_, exists := c.cursors[cursorName]
-
-	if !p.stmt.described {
-		_ = c.sendPgCursorsRowDescriptionWithFormats(p.resultFormats)
-	}
-
-	rowCount := 0
-	if exists {
-		_ = c.sendDataRowWithFormats([]interface{}{int64(1)}, p.resultFormats, []int32{23})
-		rowCount = 1
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", rowCount))
-}
-
-// sendPgCursorsRowDescription sends a RowDescription for a pg_cursors query result (single int4 column).
-func (c *clientConn) sendPgCursorsRowDescriptionWithFormats(formatCodes []int16) error {
-	var buf bytes.Buffer
-	_ = binary.Write(&buf, binary.BigEndian, int16(1)) // 1 column
-	buf.WriteString("?column?")
-	buf.WriteByte(0)
-	_ = binary.Write(&buf, binary.BigEndian, int32(0))  // table OID
-	_ = binary.Write(&buf, binary.BigEndian, int16(0))  // column attr
-	_ = binary.Write(&buf, binary.BigEndian, int32(23)) // int4 OID
-	_ = binary.Write(&buf, binary.BigEndian, int16(4))  // type size
-	_ = binary.Write(&buf, binary.BigEndian, int32(-1)) // typmod
-	var format int16
-	if len(formatCodes) == 1 {
-		format = formatCodes[0]
-	} else if len(formatCodes) > 0 {
-		format = formatCodes[0]
-	}
-	_ = binary.Write(&buf, binary.BigEndian, format)
-	return wire.WriteMessage(c.writer, wire.MsgRowDescription, buf.Bytes())
-}
-
-// pgStatActivityRegex matches queries that reference pg_stat_activity:
-//
-//	SELECT ... FROM pg_stat_activity ...
-//	SELECT ... FROM pg_catalog.pg_stat_activity ...
-//
-// Limitation: this intercepts any query containing FROM pg_stat_activity,
-// so WHERE clauses, JOINs, and column selection are ignored — all rows
-// with all columns are always returned. This is acceptable because most
-// tools simply do SELECT * FROM pg_stat_activity.
-var pgStatActivityRegex = regexp.MustCompile(
-	`(?i)\bFROM\s+(?:pg_catalog\s*\.\s*)?pg_stat_activity\b`,
-)
-
-// matchPgStatActivityQuery returns true if a query references pg_stat_activity.
-func matchPgStatActivityQuery(query string) bool {
-	if !strings.Contains(query, "pg_stat_activity") {
-		return false
-	}
-	return pgStatActivityRegex.MatchString(query)
-}
-
-// pg_stat_activity column definitions
-var pgStatActivityColumns = []struct {
-	name    string
-	oid     int32
-	typSize int16
-}{
-	{"datid", 23, 4},                 // int4
-	{"datname", 25, -1},              // text
-	{"pid", 23, 4},                   // int4
-	{"usesysid", 23, 4},              // int4
-	{"usename", 25, -1},              // text
-	{"application_name", 25, -1},     // text
-	{"client_addr", 25, -1},          // text (inet in PG, text here)
-	{"client_port", 23, 4},           // int4
-	{"backend_start", 1184, 8},       // timestamptz
-	{"xact_start", 1184, 8},          // timestamptz (NULL)
-	{"query_start", 1184, 8},         // timestamptz (NULL)
-	{"state_change", 1184, 8},        // timestamptz (NULL)
-	{"wait_event_type", 25, -1},      // text (NULL)
-	{"wait_event", 25, -1},           // text (NULL)
-	{"state", 25, -1},                // text
-	{"backend_xid", 28, 4},           // xid (NULL)
-	{"backend_xmin", 28, 4},          // xid (NULL)
-	{"query", 25, -1},                // text
-	{"backend_type", 25, -1},         // text
-	{"leader_pid", 23, 4},            // int4 (NULL)
-	{"worker_id", 23, 4},             // int4 (duckgres extension)
-	{"query_progress", 701, 8},       // float8 (percentage, -1 if not tracked)
-	{"rows_processed", 20, 8},        // int8
-	{"total_rows_to_process", 20, 8}, // int8
-}
-
-// pgStatActivityTypeOIDs is precomputed from pgStatActivityColumns to avoid per-row allocation.
-var pgStatActivityTypeOIDs = func() []int32 {
-	oids := make([]int32, len(pgStatActivityColumns))
-	for i, col := range pgStatActivityColumns {
-		oids[i] = col.oid
-	}
-	return oids
-}()
-
-// handlePgStatActivity handles SELECT FROM pg_stat_activity in the Simple Query protocol.
-func (c *clientConn) handlePgStatActivity() error {
-	if err := c.sendPgStatActivityRowDescriptionWithFormats(nil); err != nil {
-		return err
-	}
-
-	conns := c.server.listConns()
-	sort.Slice(conns, func(i, j int) bool { return conns[i].pid < conns[j].pid })
-
-	for _, conn := range conns {
-		if err := c.sendPgStatActivityDataRow(conn, nil); err != nil {
-			return err
-		}
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", len(conns)))
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handlePgStatActivityExtended handles SELECT FROM pg_stat_activity in the Extended Query protocol.
-func (c *clientConn) handlePgStatActivityExtended(p *portal) {
-	if !p.stmt.described && !p.described {
-		_ = c.sendPgStatActivityRowDescriptionWithFormats(p.resultFormats)
-	}
-
-	conns := c.server.listConns()
-	sort.Slice(conns, func(i, j int) bool { return conns[i].pid < conns[j].pid })
-
-	for _, conn := range conns {
-		_ = c.sendPgStatActivityDataRow(conn, p.resultFormats)
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("SELECT %d", len(conns)))
-}
-
-// sendPgStatActivityRowDescription sends a RowDescription for pg_stat_activity.
-func (c *clientConn) sendPgStatActivityRowDescriptionWithFormats(formatCodes []int16) error {
-	var buf bytes.Buffer
-	_ = binary.Write(&buf, binary.BigEndian, int16(len(pgStatActivityColumns)))
-	for i, col := range pgStatActivityColumns {
-		buf.WriteString(col.name)
-		buf.WriteByte(0)
-		_ = binary.Write(&buf, binary.BigEndian, int32(0))    // table OID
-		_ = binary.Write(&buf, binary.BigEndian, int16(0))    // column attr
-		_ = binary.Write(&buf, binary.BigEndian, col.oid)     // type OID
-		_ = binary.Write(&buf, binary.BigEndian, col.typSize) // type size
-		_ = binary.Write(&buf, binary.BigEndian, int32(-1))   // typmod
-		var format int16
-		if len(formatCodes) == 1 {
-			format = formatCodes[0]
-		} else if i < len(formatCodes) {
-			format = formatCodes[i]
-		}
-		_ = binary.Write(&buf, binary.BigEndian, format)
-	}
-	return wire.WriteMessage(c.writer, wire.MsgRowDescription, buf.Bytes())
-}
-
-// sendPgStatActivityDataRow sends a DataRow for a single connection in pg_stat_activity.
-func (c *clientConn) sendPgStatActivityDataRow(conn *clientConn, formatCodes []int16) error {
-	// Determine state.
-	// NOTE: conn.txStatus is read without synchronization. This is a benign race —
-	// txStatus is a single byte written only by the owning goroutine, and a stale
-	// read just means a briefly inaccurate state string. Making txStatus atomic
-	// would require changing dozens of write sites for negligible benefit.
-	var state string
-	q, _ := conn.currentQuery.Load().(string)
-	if q != "" {
-		state = "active"
-	} else {
-		switch conn.txStatus {
-		case txStatusTransaction:
-			state = "idle in transaction"
-		case txStatusError:
-			state = "idle in transaction (aborted)"
-		default:
-			state = "idle"
-		}
-	}
-
-	// Extract client IP and port from RemoteAddr
-	var clientAddr string
-	var clientPort int32
-	if conn.conn != nil {
-		if addr, ok := conn.conn.RemoteAddr().(*net.TCPAddr); ok {
-			clientAddr = addr.IP.String()
-			clientPort = int32(addr.Port)
-		}
-	}
-
-	// query_start: populated from atomic.Value when a query is active.
-	var queryStart interface{}
-	if qs, ok := conn.queryStart.Load().(time.Time); ok && !qs.IsZero() {
-		queryStart = qs
-	}
-
-	// Query progress columns: populated from cached worker health check data
-	// (control plane mode) or nil (standalone mode).
-	var queryProgress, rowsProcessed, totalRowsToProcess interface{}
-	if c.server.progressFn != nil {
-		pct, rows, total, stalled := c.server.progressFn(conn.pid)
-		queryProgress = pct
-		rowsProcessed = int64(rows)
-		totalRowsToProcess = int64(total)
-		// Override state to "active (stuck)" when the worker detects no progress.
-		if stalled && state == "active" {
-			state = "active (stuck)"
-		}
-	} else {
-		queryProgress = float64(-1)
-		rowsProcessed = int64(0)
-		totalRowsToProcess = int64(0)
-	}
-
-	values := []interface{}{
-		int32(0),             // datid
-		conn.database,        // datname
-		conn.pid,             // pid
-		int32(10),            // usesysid
-		conn.username,        // usename
-		conn.applicationName, // application_name
-		clientAddr,           // client_addr
-		clientPort,           // client_port
-		conn.backendStart,    // backend_start
-		nil,                  // xact_start (NULL)
-		queryStart,           // query_start
-		nil,                  // state_change (NULL)
-		nil,                  // wait_event_type (NULL)
-		nil,                  // wait_event (NULL)
-		state,                // state
-		nil,                  // backend_xid (NULL)
-		nil,                  // backend_xmin (NULL)
-		q,                    // query
-		"client backend",     // backend_type
-		nil,                  // leader_pid (NULL)
-		int32(conn.workerID), // worker_id
-		queryProgress,        // query_progress (float8)
-		rowsProcessed,        // rows_processed (int8)
-		totalRowsToProcess,   // total_rows_to_process (int8)
-	}
-
-	return c.sendDataRowWithFormats(values, formatCodes, pgStatActivityTypeOIDs)
-}
-
-// handleDeclareCursor handles DECLARE cursor in the Simple Query protocol.
-func (c *clientConn) handleDeclareCursor(query string, stmt *pg_query.DeclareCursorStmt) error {
-	start := time.Now()
-	innerSQL := deparseInnerQuery(stmt.Query)
-	if innerSQL == "" {
-		c.sendError("ERROR", "42601", "could not deparse cursor query")
-		c.logQuery(start, query, query, "DECLARE", 0, 0, "42601", "could not deparse cursor query", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Transpile the inner SELECT query (skip for passthrough users)
-	transpiledSQL := innerSQL
-	if !c.passthrough {
-		tr := c.newTranspiler(false)
-		result, err := tr.Transpile(innerSQL)
-		if err != nil {
-			errMsg := fmt.Sprintf("syntax error in cursor query: %v", err)
-			c.sendError("ERROR", "42601", errMsg)
-			c.logQuery(start, query, query, "DECLARE", 0, 0, "42601", errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-		transpiledSQL = result.SQL
-		if result.FallbackToNative {
-			transpiledSQL = innerSQL
-		}
-	}
-
-	// Close existing cursor with same name (PostgreSQL behavior)
-	c.closeCursor(stmt.Portalname)
-
-	c.cursors[stmt.Portalname] = &cursorState{query: transpiledSQL}
-	slog.Debug("Cursor declared.", "user", c.username, "cursor", stmt.Portalname, "query", transpiledSQL)
-
-	_ = wire.WriteCommandComplete(c.writer, "DECLARE CURSOR")
-	c.logQuery(start, query, query, "DECLARE", 0, 0, "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleFetchCursor handles FETCH in the Simple Query protocol.
-func (c *clientConn) handleFetchCursor(query string, stmt *pg_query.FetchStmt) error {
-	start := time.Now()
-	// Validate direction
-	if !isFetchForwardOnly(stmt.Direction) {
-		c.sendError("ERROR", "0A000", "cursor can only scan forward")
-		c.logQuery(start, query, query, "FETCH", 0, 0, "0A000", "cursor can only scan forward", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	cursor, ok := c.cursors[stmt.Portalname]
-	if !ok {
-		errMsg := fmt.Sprintf("cursor %q does not exist", stmt.Portalname)
-		c.sendError("ERROR", "34000", errMsg)
-		c.logQuery(start, query, query, "FETCH", 0, 0, "34000", errMsg, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Open cursor on first FETCH
-	if cursor.rows == nil {
-		if err := c.openCursor(cursor); err != nil {
-			errCode := "42000"
-			errMsg := friendlyExecError(err)
-			if c.isCallerCancellation(err) {
-				errCode = "57014"
-				errMsg = "canceling statement due to user request"
-			}
-			c.sendError("ERROR", errCode, errMsg)
-			c.setTxError()
-			c.logQuery(start, query, query, "FETCH", 0, 0, errCode, errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-	}
-
-	// Determine how many rows to fetch (pg_query sets HowMany=MaxInt64 for FETCH ALL)
-	howMany := stmt.HowMany
-	if howMany < 0 {
-		c.sendError("ERROR", "0A000", "cursor can only scan forward")
-		c.logQuery(start, query, query, "FETCH", 0, 0, "0A000", "cursor can only scan forward", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// MOVE: advance position without returning rows
-	if stmt.Ismove {
-		moveCount := int64(0)
-		for moveCount < howMany && cursor.rows.Next() {
-			// Read the row to advance position, but don't send it
-			values := make([]interface{}, len(cursor.cols))
-			valuePtrs := make([]interface{}, len(cursor.cols))
-			for i := range values {
-				valuePtrs[i] = &values[i]
-			}
-			_ = cursor.rows.Scan(valuePtrs...)
-			moveCount++
-		}
-		_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("MOVE %d", moveCount))
-		c.logQuery(start, query, query, "FETCH", moveCount, 0, "", "", "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	// Send RowDescription
-	if err := c.sendRowDescription(cursor.cols, cursor.colTypes); err != nil {
-		return err
-	}
-
-	// Stream rows
-	rowCount := int64(0)
-	for rowCount < howMany && cursor.rows.Next() {
-		values := make([]interface{}, len(cursor.cols))
-		valuePtrs := make([]interface{}, len(cursor.cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := cursor.rows.Scan(valuePtrs...); err != nil {
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			c.logQuery(start, query, query, "FETCH", 0, 0, "42000", err.Error(), "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-
-		if err := c.sendDataRowWithFormats(values, nil, cursor.typeOIDs); err != nil {
-			return err
-		}
-		rowCount++
-	}
-
-	if err := cursor.rows.Err(); err != nil {
-		errCode := "42000"
-		errMsg := friendlyExecError(err)
-		if c.isCallerCancellation(err) {
-			errCode = "57014"
-			errMsg = "canceling statement due to user request"
-		}
-		c.sendError("ERROR", errCode, errMsg)
-		c.setTxError()
-		c.logQuery(start, query, query, "FETCH", 0, 0, errCode, errMsg, "simple")
-		_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-		_ = c.writer.Flush()
-		return nil
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("FETCH %d", rowCount))
-	c.logQuery(start, query, query, "FETCH", rowCount, 0, "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleCloseCursor handles CLOSE in the Simple Query protocol.
-func (c *clientConn) handleCloseCursor(query string, stmt *pg_query.ClosePortalStmt) error {
-	start := time.Now()
-	if stmt.Portalname == "" {
-		// CLOSE ALL
-		c.closeAllCursors()
-	} else {
-		if _, ok := c.cursors[stmt.Portalname]; !ok {
-			errMsg := fmt.Sprintf("cursor %q does not exist", stmt.Portalname)
-			c.sendError("ERROR", "34000", errMsg)
-			c.logQuery(start, query, query, "CLOSE", 0, 0, "34000", errMsg, "simple")
-			_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-			_ = c.writer.Flush()
-			return nil
-		}
-		c.closeCursor(stmt.Portalname)
-	}
-
-	slog.Debug("Cursor closed.", "user", c.username, "cursor", stmt.Portalname)
-	_ = wire.WriteCommandComplete(c.writer, "CLOSE CURSOR")
-	c.logQuery(start, query, query, "CLOSE", 0, 0, "", "", "simple")
-	_ = wire.WriteReadyForQuery(c.writer, c.txStatus)
-	_ = c.writer.Flush()
-	return nil
-}
-
-// handleDeclareCursorExtended handles DECLARE cursor in the Extended Query protocol.
-func (c *clientConn) handleDeclareCursorExtended(p *portal) {
-	// Close existing cursor with same name
-	c.closeCursor(p.stmt.cursorName)
-
-	c.cursors[p.stmt.cursorName] = &cursorState{query: p.stmt.cursorQuery}
-	slog.Debug("Cursor declared (extended).", "user", c.username, "cursor", p.stmt.cursorName, "query", p.stmt.cursorQuery)
-
-	_ = wire.WriteCommandComplete(c.writer, "DECLARE CURSOR")
-}
-
-// handleFetchCursorExtended handles FETCH in the Extended Query protocol.
-func (c *clientConn) handleFetchCursorExtended(p *portal) {
-	cursor, ok := c.cursors[p.stmt.cursorName]
-	if !ok {
-		c.sendError("ERROR", "34000", fmt.Sprintf("cursor %q does not exist", p.stmt.cursorName))
-		return
-	}
-
-	// Open cursor on first FETCH
-	if cursor.rows == nil {
-		if err := c.openCursor(cursor); err != nil {
-			if c.isCallerCancellation(err) {
-				c.sendError("ERROR", "57014", "canceling statement due to user request")
-			} else {
-				c.sendError("ERROR", "42000", err.Error())
-			}
-			c.setTxError()
-			return
-		}
-	}
-
-	howMany := p.stmt.fetchCount
-
-	// Send RowDescription if Describe wasn't already called
-	if !p.described && len(cursor.cols) > 0 {
-		if err := c.sendRowDescriptionWithFormats(cursor.cols, cursor.colTypes, p.resultFormats); err != nil {
-			return
-		}
-	}
-
-	// Stream rows
-	rowCount := int64(0)
-	for rowCount < howMany && cursor.rows.Next() {
-		values := make([]interface{}, len(cursor.cols))
-		valuePtrs := make([]interface{}, len(cursor.cols))
-		for i := range values {
-			valuePtrs[i] = &values[i]
-		}
-
-		if err := cursor.rows.Scan(valuePtrs...); err != nil {
-			c.sendError("ERROR", "42000", err.Error())
-			c.setTxError()
-			return
-		}
-
-		if err := c.sendDataRowWithFormats(values, p.resultFormats, cursor.typeOIDs); err != nil {
-			return
-		}
-		rowCount++
-	}
-
-	if err := cursor.rows.Err(); err != nil {
-		if c.isCallerCancellation(err) {
-			c.sendError("ERROR", "57014", "canceling statement due to user request")
-		} else {
-			c.sendError("ERROR", "42000", err.Error())
-		}
-		c.setTxError()
-		return
-	}
-
-	_ = wire.WriteCommandComplete(c.writer, fmt.Sprintf("FETCH %d", rowCount))
-}
-
-// handleCloseCursorExtended handles CLOSE cursor in the Extended Query protocol.
-func (c *clientConn) handleCloseCursorExtended(p *portal) {
-	if p.stmt.cursorName == "" {
-		c.closeAllCursors()
-	} else {
-		if _, ok := c.cursors[p.stmt.cursorName]; !ok {
-			c.sendError("ERROR", "34000", fmt.Sprintf("cursor %q does not exist", p.stmt.cursorName))
-			return
-		}
-		c.closeCursor(p.stmt.cursorName)
-	}
-
-	slog.Debug("Cursor closed (extended).", "user", c.username, "cursor", p.stmt.cursorName)
-	_ = wire.WriteCommandComplete(c.writer, "CLOSE CURSOR")
-}
-
-// isAlterTableNotTableError checks if the error indicates that an ALTER TABLE
-// was attempted on a view. DuckDB returns this error when trying to use
-// ALTER TABLE ... RENAME TO on a view instead of ALTER VIEW.
-func isAlterTableNotTableError(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	msg := strings.ToLower(err.Error())
-	if strings.Contains(msg, "cannot use alter table") &&
-		strings.Contains(msg, "not a table") {
-		return true
-	}
-
-	if strings.Contains(msg, "can only modify view with alter view statement") {
-		return true
-	}
-
-	return false
-}
-
-// isDropTableOnViewError checks if the error indicates that a DROP TABLE
-// was attempted on a view. DuckDB returns:
-// "Catalog Error: Existing object X is of type View, trying to drop type Table"
-func isDropTableOnViewError(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	msg := err.Error()
-	return strings.Contains(msg, "is of type View") &&
-		strings.Contains(msg, "trying to drop type Table")
 }
